Network.AWS.Support

Description

AWS Support is a one-on-one, fast-response support channel that is staffed with experienced support personnel. The service helps customers get the most from the products and features provided by Amazon Web Services. There are four levels, or tiers, of AWS Support: Basic, Developer, Business, and Enterprise. The Basic tier is free of charge and offers support for account and billing questions and service limit increases. The other tiers offer an unlimited number of technical support cases with pay-by-the-month pricing and no long-term contracts, providing developers and businesses flexibility to choose the level of support that meets their needs.

Documentation

module Network.AWS.Support.AddAttachmentsToSet

module Network.AWS.Support.AddCommunicationToCase

module Network.AWS.Support.CreateCase

module Network.AWS.Support.DescribeAttachment

module Network.AWS.Support.DescribeCases

module Network.AWS.Support.DescribeCommunications

module Network.AWS.Support.DescribeServices

module Network.AWS.Support.DescribeSeverityLevels

module Network.AWS.Support.DescribeTrustedAdvisorCheckRefreshStatuses

module Network.AWS.Support.DescribeTrustedAdvisorCheckResult

module Network.AWS.Support.DescribeTrustedAdvisorCheckSummaries

module Network.AWS.Support.DescribeTrustedAdvisorChecks

module Network.AWS.Support.RefreshTrustedAdvisorCheck

module Network.AWS.Support.ResolveCase

module Network.AWS.Support.Types