Network.AWS.Support.CreateCase

Description

Creates a new case in the AWS Support Center. This operation is modeled on the behavior of the AWS Support Center Create Case page. Its parameters require you to specify the following information:

IssueType. The type of issue for the case. You can specify either "customer-service" or "technical." If you do not indicate a value, the default is "technical." ServiceCode. The code for an AWS service. You obtain the ServiceCode by calling DescribeServices. CategoryCode. The category for the service defined for the ServiceCode value. You also obtain the category code for a service by calling DescribeServices. Each AWS service defines its own set of category codes. SeverityCode. A value that indicates the urgency of the case, which in turn determines the response time according to your service level agreement with AWS Support. You obtain the SeverityCode by calling DescribeSeverityLevels. Subject. The Subject field on the AWS Support Center Create Case page. CommunicationBody. The Description field on the AWS Support Center Create Case page. AttachmentSetId. The ID of a set of attachments that has been created by using AddAttachmentsToSet. Language. The human language in which AWS Support handles the case. English and Japanese are currently supported. CcEmailAddresses. The AWS Support Center CC field on the Create Case page. You can list email addresses to be copied on any correspondence about the case. The account that opens the case is already identified by passing the AWS Credentials in the HTTP POST method or in a method or function call from one of the programming languages supported by an AWS SDK. To add additional communication or attachments to an existing case, use AddCommunicationToCase.

A successful CreateCase request returns an AWS Support case number. Case numbers are used by the DescribeCases operation to retrieve existing AWS Support cases.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_CreateCase.html

Request

data CreateCase

Request constructor

createCase

Arguments

:: Text
ccSubject

-> Text
ccCommunicationBody

-> CreateCase

CreateCase constructor.

The fields accessible through corresponding lenses are:

ccAttachmentSetId :: Maybe Text

ccCategoryCode :: Maybe Text

ccCcEmailAddresses :: [Text]

ccCommunicationBody :: Text

ccIssueType :: Maybe Text

ccLanguage :: Maybe Text

ccServiceCode :: Maybe Text

ccSeverityCode :: Maybe Text

ccSubject :: Text

Request lenses

ccAttachmentSetId :: Lens' CreateCase (Maybe Text)

The ID of a set of one or more attachments for the case. Create the set by using AddAttachmentsToSet.

ccCategoryCode :: Lens' CreateCase (Maybe Text)

The category of problem for the AWS Support case.

ccCcEmailAddresses :: Lens' CreateCase [Text]

A list of email addresses that AWS Support copies on case correspondence.

ccCommunicationBody :: Lens' CreateCase Text

The communication body text when you create an AWS Support case by calling CreateCase.

ccIssueType :: Lens' CreateCase (Maybe Text)

The type of issue for the case. You can specify either "customer-service" or "technical." If you do not indicate a value, the default is "technical."

ccLanguage :: Lens' CreateCase (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

ccServiceCode :: Lens' CreateCase (Maybe Text)

The code for the AWS service returned by the call to DescribeServices.

ccSeverityCode :: Lens' CreateCase (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels .

The availability of severity levels depends on each customer's support subscription. In other words, your subscription may not necessarily require the urgent level of response time.

ccSubject :: Lens' CreateCase Text

The title of the AWS Support case.

Response

data CreateCaseResponse

Response constructor

createCaseResponse :: CreateCaseResponse

CreateCaseResponse constructor.

The fields accessible through corresponding lenses are:

ccrCaseId :: Maybe Text

Response lenses

ccrCaseId :: Lens' CreateCaseResponse (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47