Network.AWS.Support.DescribeAttachment

Description

Returns the attachment that has the specified ID. Attachment IDs are generated by the case management system when you add an attachment to a case or case communication. Attachment IDs are returned in the AttachmentDetails objects that are returned by the DescribeCommunications operation.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_DescribeAttachment.html

Request

data DescribeAttachment

Request constructor

describeAttachment

Arguments

:: Text
daAttachmentId

-> DescribeAttachment

DescribeAttachment constructor.

The fields accessible through corresponding lenses are:

daAttachmentId :: Text

Request lenses

daAttachmentId :: Lens' DescribeAttachment Text

The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

Response

data DescribeAttachmentResponse

Response constructor

describeAttachmentResponse :: DescribeAttachmentResponse

DescribeAttachmentResponse constructor.

The fields accessible through corresponding lenses are:

darAttachment :: Maybe Attachment

Response lenses

darAttachment :: Lens' DescribeAttachmentResponse (Maybe Attachment)

The attachment content and file name.