Network.AWS.Support.DescribeCases

Description

Returns a list of cases that you specify by passing one or more case IDs. In addition, you can filter the cases by date by setting values for the AfterTime and BeforeTime request parameters. You can set values for the IncludeResolvedCases and IncludeCommunications request parameters to control how much information is returned.

Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error.

The response returns the following in JSON format:

One or more CaseDetails data types. One or more NextToken values, which specify where to paginate the returned records represented by the CaseDetails objects.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_DescribeCases.html

Request

data DescribeCases

Request constructor

describeCases :: DescribeCases

DescribeCases constructor.

The fields accessible through corresponding lenses are:

dcAfterTime :: Maybe Text

dcBeforeTime :: Maybe Text

dcCaseIdList :: [Text]

dcDisplayId :: Maybe Text

dcIncludeCommunications :: Maybe Bool

dcIncludeResolvedCases :: Maybe Bool

dcLanguage :: Maybe Text

dcMaxResults :: Maybe Natural

dcNextToken :: Maybe Text

Request lenses

dcAfterTime :: Lens' DescribeCases (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dcBeforeTime :: Lens' DescribeCases (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dcCaseIdList :: Lens' DescribeCases [Text]

A list of ID numbers of the support cases you want returned. The maximum number of cases is 100.

dcDisplayId :: Lens' DescribeCases (Maybe Text)

The ID displayed for a case in the AWS Support Center user interface.

dcIncludeCommunications :: Lens' DescribeCases (Maybe Bool)

Specifies whether communications should be included in the DescribeCases results. The default is true.

dcIncludeResolvedCases :: Lens' DescribeCases (Maybe Bool)

Specifies whether resolved support cases should be included in the DescribeCases results. The default is false.

dcLanguage :: Lens' DescribeCases (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

dcMaxResults :: Lens' DescribeCases (Maybe Natural)

The maximum number of results to return before paginating.

dcNextToken :: Lens' DescribeCases (Maybe Text)

A resumption point for pagination.

Response

data DescribeCasesResponse

Response constructor

describeCasesResponse :: DescribeCasesResponse

DescribeCasesResponse constructor.

The fields accessible through corresponding lenses are:

dcr1Cases :: [CaseDetails]

dcr1NextToken :: Maybe Text

Response lenses

dcr1Cases :: Lens' DescribeCasesResponse [CaseDetails]

The details for the cases that match the request.

dcr1NextToken :: Lens' DescribeCasesResponse (Maybe Text)

A resumption point for pagination.