Network.AWS.Support.DescribeCommunications

Description

Returns communications (and attachments) for one or more support cases. You can use the AfterTime and BeforeTime parameters to filter by date. You can use the CaseId parameter to restrict the results to a particular case.

Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error.

You can use the MaxResults and NextToken parameters to control the pagination of the result set. Set MaxResults to the number of cases you want displayed on each page, and use NextToken to specify the resumption of pagination.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_DescribeCommunications.html

Request

data DescribeCommunications

Request constructor

describeCommunications

Arguments

:: Text
dc1CaseId

-> DescribeCommunications

DescribeCommunications constructor.

The fields accessible through corresponding lenses are:

dc1AfterTime :: Maybe Text

dc1BeforeTime :: Maybe Text

dc1CaseId :: Text

dc1MaxResults :: Maybe Natural

dc1NextToken :: Maybe Text

Request lenses

dc1AfterTime :: Lens' DescribeCommunications (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dc1BeforeTime :: Lens' DescribeCommunications (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dc1CaseId :: Lens' DescribeCommunications Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

dc1MaxResults :: Lens' DescribeCommunications (Maybe Natural)

The maximum number of results to return before paginating.

dc1NextToken :: Lens' DescribeCommunications (Maybe Text)

A resumption point for pagination.

Response

data DescribeCommunicationsResponse

Response constructor

describeCommunicationsResponse :: DescribeCommunicationsResponse

DescribeCommunicationsResponse constructor.

The fields accessible through corresponding lenses are:

dcrCommunications :: [Communication]

dcrNextToken :: Maybe Text

Response lenses

dcrCommunications :: Lens' DescribeCommunicationsResponse [Communication]

The communications for the case.

dcrNextToken :: Lens' DescribeCommunicationsResponse (Maybe Text)

A resumption point for pagination.