Network.AWS.Support.DescribeSeverityLevels

Description

Returns the list of severity levels that you can assign to an AWS Support case. The severity level for a case is also a field in the CaseDetails data type included in any CreateCase request.

Creating a Request

describeSeverityLevels :: DescribeSeverityLevels

Creates a value of DescribeSeverityLevels with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dslLanguage - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

data DescribeSeverityLevels

See: describeSeverityLevels smart constructor.

Request Lenses

dslLanguage :: Lens' DescribeSeverityLevels (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Destructuring the Response

describeSeverityLevelsResponse

Arguments

:: Int
dslrsResponseStatus

-> DescribeSeverityLevelsResponse

Creates a value of DescribeSeverityLevelsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dslrsSeverityLevels - The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

dslrsResponseStatus - -- | The response status code.

data DescribeSeverityLevelsResponse

The list of severity levels returned by the DescribeSeverityLevels operation.

See: describeSeverityLevelsResponse smart constructor.

Response Lenses

dslrsSeverityLevels :: Lens' DescribeSeverityLevelsResponse [SeverityLevel]

The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

dslrsResponseStatus :: Lens' DescribeSeverityLevelsResponse Int

- | The response status code.