Network.AWS.Support.DescribeTrustedAdvisorCheckRefreshStatuses

Description

Returns the refresh status of the Trusted Advisor checks that have the specified check IDs. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_DescribeTrustedAdvisorCheckRefreshStatuses.html

Request

data DescribeTrustedAdvisorCheckRefreshStatuses

Request constructor

describeTrustedAdvisorCheckRefreshStatuses :: DescribeTrustedAdvisorCheckRefreshStatuses

DescribeTrustedAdvisorCheckRefreshStatuses constructor.

The fields accessible through corresponding lenses are:

dtacrsCheckIds :: [Text]

Request lenses

dtacrsCheckIds :: Lens' DescribeTrustedAdvisorCheckRefreshStatuses [Text]

The IDs of the Trusted Advisor checks.

Response

data DescribeTrustedAdvisorCheckRefreshStatusesResponse

Response constructor

describeTrustedAdvisorCheckRefreshStatusesResponse :: DescribeTrustedAdvisorCheckRefreshStatusesResponse

DescribeTrustedAdvisorCheckRefreshStatusesResponse constructor.

The fields accessible through corresponding lenses are:

dtacrsrStatuses :: [TrustedAdvisorCheckRefreshStatus]

Response lenses

dtacrsrStatuses :: Lens' DescribeTrustedAdvisorCheckRefreshStatusesResponse [TrustedAdvisorCheckRefreshStatus]

The refresh status of the specified Trusted Advisor checks.