Network.AWS.Support.DescribeTrustedAdvisorCheckSummaries

Description

Returns the summaries of the results of the Trusted Advisor checks that have the specified check IDs. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks.

The response contains an array of TrustedAdvisorCheckSummary objects.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_DescribeTrustedAdvisorCheckSummaries.html

Request

data DescribeTrustedAdvisorCheckSummaries

Request constructor

describeTrustedAdvisorCheckSummaries :: DescribeTrustedAdvisorCheckSummaries

DescribeTrustedAdvisorCheckSummaries constructor.

The fields accessible through corresponding lenses are:

dtacsCheckIds :: [Text]

Request lenses

dtacsCheckIds :: Lens' DescribeTrustedAdvisorCheckSummaries [Text]

The IDs of the Trusted Advisor checks.

Response

data DescribeTrustedAdvisorCheckSummariesResponse

Response constructor

describeTrustedAdvisorCheckSummariesResponse :: DescribeTrustedAdvisorCheckSummariesResponse

DescribeTrustedAdvisorCheckSummariesResponse constructor.

The fields accessible through corresponding lenses are:

dtacsrSummaries :: [TrustedAdvisorCheckSummary]

Response lenses

dtacsrSummaries :: Lens' DescribeTrustedAdvisorCheckSummariesResponse [TrustedAdvisorCheckSummary]

The summary information for the requested Trusted Advisor checks.