Network.AWS.Support.DescribeTrustedAdvisorChecks

Description

Returns information about all available Trusted Advisor checks, including name, ID, category, description, and metadata. You must specify a language code; English ("en") and Japanese ("ja") are currently supported. The response contains a TrustedAdvisorCheckDescription for each check.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_DescribeTrustedAdvisorChecks.html

Request

data DescribeTrustedAdvisorChecks

Request constructor

describeTrustedAdvisorChecks

Arguments

:: Text
dtacLanguage

-> DescribeTrustedAdvisorChecks

DescribeTrustedAdvisorChecks constructor.

The fields accessible through corresponding lenses are:

dtacLanguage :: Text

Request lenses

dtacLanguage :: Lens' DescribeTrustedAdvisorChecks Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Response

data DescribeTrustedAdvisorChecksResponse

Response constructor

describeTrustedAdvisorChecksResponse :: DescribeTrustedAdvisorChecksResponse

DescribeTrustedAdvisorChecksResponse constructor.

The fields accessible through corresponding lenses are:

dtacrChecks :: [TrustedAdvisorCheckDescription]

Response lenses

dtacrChecks :: Lens' DescribeTrustedAdvisorChecksResponse [TrustedAdvisorCheckDescription]

Information about all available Trusted Advisor checks.