Network.AWS.Support.Monadic

Description

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.Support with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.Support

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.Support.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Control.Applicative
import Network.AWS.Support.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

-- Or to void any additional parameters outside of those required using empty:
operationName w x empty

AddAttachmentsToSet

Adds one or more attachments to an attachment set. If an AttachmentSetId is not specified, a new attachment set is created, and the ID of the set is returned in the response. If an AttachmentSetId is specified, the attachments are added to the specified set, if it exists. An attachment set is a temporary container for attachments that are to be added to a case or case communication. The set is available for one hour after it is created; the ExpiryTime returned in the response indicates when the set expires. The maximum number of attachments in a set is 3, and the maximum size of any attachment in the set is 5 MB.

See: AddAttachmentsToSet

addAttachmentsToSet

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Attachment]
aatsAttachments

-> State AddAttachmentsToSet a

-> m AddAttachmentsToSetResponse

addAttachmentsToSetCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Attachment]
aatsAttachments

-> State AddAttachmentsToSet a

-> m (Either ServiceEr AddAttachmentsToSetResponse)

AddCommunicationToCase

Adds additional customer communication to an AWS Support case. You use the CaseId value to identify the case to add communication to. You can list a set of email addresses to copy on the communication using the CcEmailAddresses value. The CommunicationBody value contains the text of the communication. The response indicates the success or failure of the request. This operation implements a subset of the behavior on the AWS Support Your Support Cases web form.

See: AddCommunicationToCase

addCommunicationToCase

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
actcCommunicationBody

-> State AddCommunicationToCase a

-> m AddCommunicationToCaseResponse

addCommunicationToCaseCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
actcCommunicationBody

-> State AddCommunicationToCase a

-> m (Either ServiceEr AddCommunicationToCaseResponse)

CreateCase

Creates a new case in the AWS Support Center. This operation is modeled on the behavior of the AWS Support Center Open a new case page. Its parameters require you to specify the following information: IssueType. The type of issue for the case. You can specify either "customer-service" or "technical." If you do not indicate a value, the default is "technical." ServiceCode. The code for an AWS service. You obtain the ServiceCode by calling DescribeServices. CategoryCode. The category for the service defined for the ServiceCode value. You also obtain the category code for a service by calling DescribeServices. Each AWS service defines its own set of category codes. SeverityCode. A value that indicates the urgency of the case, which in turn determines the response time according to your service level agreement with AWS Support. You obtain the SeverityCode by calling DescribeSeverityLevels. Subject. The Subject field on the AWS Support Center Open a new case page. CommunicationBody. The Description field on the AWS Support Center Open a new case page. AttachmentSetId. The ID of a set of attachments that has been created by using AddAttachmentsToSet. Language. The human language in which AWS Support handles the case. English and Japanese are currently supported. CcEmailAddresses. The AWS Support Center CC field on the Open a new case page. You can list email addresses to be copied on any correspondence about the case. The account that opens the case is already identified by passing the AWS Credentials in the HTTP POST method or in a method or function call from one of the programming languages supported by an AWS SDK. To add additional communication or attachments to an existing case, use AddCommunicationToCase. A successful CreateCase request returns an AWS Support case number. Case numbers are used by the DescribeCases operation to retrieve existing AWS Support cases.

See: CreateCase

createCase

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ccSubject

-> Text
ccCommunicationBody

-> State CreateCase a

-> m CreateCaseResponse

createCaseCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ccSubject

-> Text
ccCommunicationBody

-> State CreateCase a

-> m (Either ServiceEr CreateCaseResponse)

DescribeAttachment

Returns the attachment that has the specified ID. Attachment IDs are generated by the case management system when you add an attachment to a case or case communication. Attachment IDs are returned in the AttachmentDetails objects that are returned by the DescribeCommunications operation.

See: DescribeAttachment

describeAttachment

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
daAttachmentId

-> m DescribeAttachmentResponse

describeAttachmentCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
daAttachmentId

-> m (Either ServiceEr DescribeAttachmentResponse)

DescribeCases

Returns a list of cases that you specify by passing one or more case IDs. In addition, you can filter the cases by date by setting values for the AfterTime and BeforeTime request parameters. Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error. The response returns the following in JSON format: One or more CaseDetails data types. One or more NextToken values, which specify where to paginate the returned records represented by the CaseDetails objects.

See: DescribeCases

describeCases :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State DescribeCases a -> ResumableSource m DescribeCasesResponse

describeCasesCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State DescribeCases a -> ResumableSource m (Either ServiceEr DescribeCasesResponse)

DescribeCommunications

Returns communications (and attachments) for one or more support cases. You can use the AfterTime and BeforeTime parameters to filter by date. You can use the CaseId parameter to restrict the results to a particular case. Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error. You can use the MaxResults and NextToken parameters to control the pagination of the result set. Set MaxResults to the number of cases you want displayed on each page, and use NextToken to specify the resumption of pagination.

See: DescribeCommunications

describeCommunications

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> Text
dc1CaseId

-> State DescribeCommunications a

-> ResumableSource m DescribeCommunicationsResponse

describeCommunicationsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> Text
dc1CaseId

-> State DescribeCommunications a

-> ResumableSource m (Either ServiceEr DescribeCommunicationsResponse)

DescribeServices

Returns the current list of AWS services and a list of service categories that applies to each one. You then use service names and categories in your CreateCase requests. Each AWS service has its own set of categories. The service codes and category codes correspond to the values that are displayed in the Service and Category drop-down lists on the AWS Support Center Open a new case page. The values in those fields, however, do not necessarily match the service codes and categories returned by the DescribeServices request. Always use the service codes and categories obtained programmatically. This practice ensures that you always have the most recent set of service and category codes.

See: DescribeServices

describeServices :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeServices a -> m DescribeServicesResponse

describeServicesCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeServices a -> m (Either ServiceEr DescribeServicesResponse)

DescribeSeverityLevels

Returns the list of severity levels that you can assign to an AWS Support case. The severity level for a case is also a field in the CaseDetails data type included in any CreateCase request.

See: DescribeSeverityLevels

describeSeverityLevels :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeSeverityLevels a -> m DescribeSeverityLevelsResponse

describeSeverityLevelsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeSeverityLevels a -> m (Either ServiceEr DescribeSeverityLevelsResponse)

DescribeTrustedAdvisorCheckRefreshStatuses

Returns the refresh status of the Trusted Advisor checks that have the specified check IDs. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks.

See: DescribeTrustedAdvisorCheckRefreshStatuses

describeTrustedAdvisorCheckRefreshStatuses

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
dtacrsCheckIds

-> m DescribeTrustedAdvisorCheckRefreshStatusesResponse

describeTrustedAdvisorCheckRefreshStatusesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
dtacrsCheckIds

-> m (Either ServiceEr DescribeTrustedAdvisorCheckRefreshStatusesResponse)

DescribeTrustedAdvisorCheckResult

Returns the results of the Trusted Advisor check that has the specified check ID. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks. The response contains a TrustedAdvisorCheckResult object, which contains these three objects: TrustedAdvisorCategorySpecificSummary TrustedAdvisorResourceDetail TrustedAdvisorResourcesSummary In addition, the response contains these fields: Status. The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available". Timestamp. The time of the last refresh of the check. CheckId. The unique identifier for the check.

See: DescribeTrustedAdvisorCheckResult

describeTrustedAdvisorCheckResult

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dtacrCheckId

-> State DescribeTrustedAdvisorCheckResult a

-> m DescribeTrustedAdvisorCheckResultResponse

describeTrustedAdvisorCheckResultCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dtacrCheckId

-> State DescribeTrustedAdvisorCheckResult a

-> m (Either ServiceEr DescribeTrustedAdvisorCheckResultResponse)

DescribeTrustedAdvisorCheckSummaries

Returns the summaries of the results of the Trusted Advisor checks that have the specified check IDs. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks. The response contains an array of TrustedAdvisorCheckSummary objects.

See: DescribeTrustedAdvisorCheckSummaries

describeTrustedAdvisorCheckSummaries

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
dtacsCheckIds

-> m DescribeTrustedAdvisorCheckSummariesResponse

describeTrustedAdvisorCheckSummariesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
dtacsCheckIds

-> m (Either ServiceEr DescribeTrustedAdvisorCheckSummariesResponse)

DescribeTrustedAdvisorChecks

Returns information about all available Trusted Advisor checks, including name, ID, category, description, and metadata. You must specify a language code; English ("en") and Japanese ("ja") are currently supported. The response contains a TrustedAdvisorCheckDescription for each check.

See: DescribeTrustedAdvisorChecks

describeTrustedAdvisorChecks

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dtacLanguage

-> m DescribeTrustedAdvisorChecksResponse

describeTrustedAdvisorChecksCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dtacLanguage

-> m (Either ServiceEr DescribeTrustedAdvisorChecksResponse)

RefreshTrustedAdvisorCheck

Requests a refresh of the Trusted Advisor check that has the specified check ID. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks. The response contains a RefreshTrustedAdvisorCheckResult object, which contains these fields: Status. The refresh status of the check: "none", "enqueued", "processing", "success", or "abandoned". MillisUntilNextRefreshable. The amount of time, in milliseconds, until the check is eligible for refresh. CheckId. The unique identifier for the check.

See: RefreshTrustedAdvisorCheck

refreshTrustedAdvisorCheck

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
rtacCheckId

-> m RefreshTrustedAdvisorCheckResponse

refreshTrustedAdvisorCheckCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
rtacCheckId

-> m (Either ServiceEr RefreshTrustedAdvisorCheckResponse)

ResolveCase

Takes a CaseId and returns the initial state of the case along with the state of the case after the call to ResolveCase completed.

See: ResolveCase

resolveCase :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State ResolveCase a -> m ResolveCaseResponse

resolveCaseCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State ResolveCase a -> m (Either ServiceEr ResolveCaseResponse)

Re-exported

module Network.AWS.Support

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()