Network.AWS.Support.ResolveCase

Description

Takes a CaseId and returns the initial state of the case along with the state of the case after the call to ResolveCase completed.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_ResolveCase.html

Request

data ResolveCase

Request constructor

resolveCase :: ResolveCase

ResolveCase constructor.

The fields accessible through corresponding lenses are:

rcCaseId :: Maybe Text

Request lenses

rcCaseId :: Lens' ResolveCase (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

Response

data ResolveCaseResponse

Response constructor

resolveCaseResponse :: ResolveCaseResponse

ResolveCaseResponse constructor.

The fields accessible through corresponding lenses are:

rcrFinalCaseStatus :: Maybe Text

rcrInitialCaseStatus :: Maybe Text

Response lenses

rcrFinalCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)

The status of the case after the ResolveCase request was processed.

rcrInitialCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)

The status of the case when the ResolveCase request was sent.