Network.AWS.Support.V2013_04_15.AddAttachmentsToSet

Description

Adds one or more attachments to an attachment set. If an AttachmentSetId is not specified, a new attachment set is created, and the ID of the set is returned in the response. If an AttachmentSetId is specified, the attachments are added to the specified set, if it exists. An attachment set is a temporary container for attachments that are to be added to a case or case communication. The set is available for one hour after it is created; the ExpiryTime returned in the response indicates when the set expires. The maximum number of attachments in a set is 3, and the maximum size of any attachment in the set is 5 MB.

Request

data AddAttachmentsToSet

Request constructor

mkAddAttachmentsToSet

Arguments

:: [Attachment]
aatsAttachments

-> AddAttachmentsToSet

Smart constructor for the minimum required parameters to construct a valid AddAttachmentsToSet request.

Request lenses

aatsAttachmentSetId :: Lens' AddAttachmentsToSet (Maybe Text)

The ID of the attachment set. If an AttachmentSetId is not specified, a new attachment set is created, and the ID of the set is returned in the response. If an AttachmentSetId is specified, the attachments are added to the specified set, if it exists.

aatsAttachments :: Lens' AddAttachmentsToSet [Attachment]

One or more attachments to add to the set. The limit is 3 attachments per set, and the size limit is 5 MB per attachment.

Response

data AddAttachmentsToSetResponse

The ID and expiry time of the attachment set returned by the AddAttachmentsToSet operation.

Response lenses

aatsrAttachmentSetId :: Lens' AddAttachmentsToSetResponse (Maybe Text)

The ID of the attachment set. If an AttachmentSetId was not specified, a new attachment set is created, and the ID of the set is returned in the response. If an AttachmentSetId was specified, the attachments are added to the specified set, if it exists.

aatsrExpiryTime :: Lens' AddAttachmentsToSetResponse (Maybe Text)

The time and date when the attachment set expires.