Network.AWS.Support.V2013_04_15.AddCommunicationToCase

Description

Adds additional customer communication to an AWS Support case. You use the CaseId value to identify the case to add communication to. You can list a set of email addresses to copy on the communication using the CcEmailAddresses value. The CommunicationBody value contains the text of the communication. The response indicates the success or failure of the request. This operation implements a subset of the behavior on the AWS Support Your Support Cases web form.

Documentation

addCommunicationToCase

Arguments

:: Text
_actcrCommunicationBody

-> AddCommunicationToCase

Minimum specification for a AddCommunicationToCase request.

data AddCommunicationToCase

Constructors

AddCommunicationToCase

Fields

_actcrCommunicationBody :: Text

The body of an email communication to add to the support case.

_actcrAttachmentSetId :: Maybe Text

The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet.

_actcrCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_actcrCcEmailAddresses :: [Text]

The email addresses in the CC line of an email to be added to the support case.

actcrCommunicationBody :: Lens' AddCommunicationToCase Text

actcrCcEmailAddresses :: Lens' AddCommunicationToCase [Text]

actcrCaseId :: Lens' AddCommunicationToCase (Maybe Text)

actcrAttachmentSetId :: Lens' AddCommunicationToCase (Maybe Text)

data AddCommunicationToCaseResponse

Constructors

AddCommunicationToCaseResponse

Fields

_actcsResult :: Maybe Bool

True if AddCommunicationToCase succeeds. Otherwise, returns an error.

actcsResult :: Iso' AddCommunicationToCaseResponse (Maybe Bool)