Network.AWS.Support.V2013_04_15.DescribeCases

Description

Returns a list of cases that you specify by passing one or more case IDs. In addition, you can filter the cases by date by setting values for the AfterTime and BeforeTime request parameters. Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error. The response returns the following in JSON format: One or more CaseDetails data types. One or more NextToken values, which specify where to paginate the returned records represented by the CaseDetails objects.

Documentation

describeCases :: DescribeCases

Minimum specification for a DescribeCases request.

data DescribeCases

Constructors

DescribeCases

Fields

_dcrAfterTime :: Maybe Text

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dcrBeforeTime :: Maybe Text

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dcrCaseIdList :: [Text]

A list of ID numbers of the support cases you want returned. The maximum number of cases is 100.

_dcrDisplayId :: Maybe Text

The ID displayed for a case in the AWS Support Center user interface.

_dcrIncludeCommunications :: Maybe Bool

Specifies whether communications should be included in the DescribeCases results. The default is true.

_dcrIncludeResolvedCases :: Maybe Bool

Specifies whether resolved support cases should be included in the DescribeCases results. The default is false.

_dcrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

_dcrMaxResults :: Maybe Integer

The maximum number of results to return before paginating.

_dcrNextToken :: Maybe Text

A resumption point for pagination.

dcrNextToken :: Lens' DescribeCases (Maybe Text)

dcrMaxResults :: Lens' DescribeCases (Maybe Integer)

dcrLanguage :: Lens' DescribeCases (Maybe Text)

dcrIncludeResolvedCases :: Lens' DescribeCases (Maybe Bool)

dcrIncludeCommunications :: Lens' DescribeCases (Maybe Bool)

dcrDisplayId :: Lens' DescribeCases (Maybe Text)

dcrCaseIdList :: Lens' DescribeCases [Text]

dcrBeforeTime :: Lens' DescribeCases (Maybe Text)

dcrAfterTime :: Lens' DescribeCases (Maybe Text)

data DescribeCasesResponse

Constructors

DescribeCasesResponse

Fields

_dcsCases :: [CaseDetails]

The details for the cases that match the request.

_dcsNextToken :: Maybe Text

A resumption point for pagination.

dcsNextToken :: Lens' DescribeCasesResponse (Maybe Text)

dcsCases :: Lens' DescribeCasesResponse [CaseDetails]