Network.AWS.Support.V2013_04_15.DescribeCommunications

Description

Returns communications (and attachments) for one or more support cases. You can use the AfterTime and BeforeTime parameters to filter by date. You can use the CaseId parameter to restrict the results to a particular case. Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error. You can use the MaxResults and NextToken parameters to control the pagination of the result set. Set MaxResults to the number of cases you want displayed on each page, and use NextToken to specify the resumption of pagination.

Documentation

describeCommunications

Arguments

:: Text
_dctCaseId

-> DescribeCommunications

Minimum specification for a DescribeCommunications request.

data DescribeCommunications

Constructors

DescribeCommunications

Fields

_dctCaseId :: Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_dctAfterTime :: Maybe Text

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dctBeforeTime :: Maybe Text

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dctMaxResults :: Maybe Integer

The maximum number of results to return before paginating.

_dctNextToken :: Maybe Text

A resumption point for pagination.

dctNextToken :: Lens' DescribeCommunications (Maybe Text)

dctMaxResults :: Lens' DescribeCommunications (Maybe Integer)

dctCaseId :: Lens' DescribeCommunications Text

dctBeforeTime :: Lens' DescribeCommunications (Maybe Text)

dctAfterTime :: Lens' DescribeCommunications (Maybe Text)

data DescribeCommunicationsResponse

Constructors

DescribeCommunicationsResponse

Fields

_dcuCommunications :: [Communication]

The communications for the case.

_dcuNextToken :: Maybe Text

A resumption point for pagination.

dcuNextToken :: Lens' DescribeCommunicationsResponse (Maybe Text)

dcuCommunications :: Lens' DescribeCommunicationsResponse [Communication]