Network.AWS.Support.V2013_04_15.DescribeServices

Description

Returns the current list of AWS services and a list of service categories that applies to each one. You then use service names and categories in your CreateCase requests. Each AWS service has its own set of categories. The service codes and category codes correspond to the values that are displayed in the Service and Category drop-down lists on the AWS Support Center Open a new case page. The values in those fields, however, do not necessarily match the service codes and categories returned by the DescribeServices request. Always use the service codes and categories obtained programmatically. This practice ensures that you always have the most recent set of service and category codes.

Request

data DescribeServices

Request constructor

mkDescribeServices :: DescribeServices

Smart constructor for the minimum required parameters to construct a valid DescribeServices request.

Request lenses

dsServiceCodeList :: Lens' DescribeServices [Text]

A JSON-formatted list of service codes available for AWS services.

dsLanguage :: Lens' DescribeServices (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Response

data DescribeServicesResponse

The list of AWS services returned by the DescribeServices operation.

Response constructor

mkDescribeServicesResponse :: DescribeServicesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeServicesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dsrServices :: Lens' DescribeServicesResponse [Service]

A JSON-formatted list of AWS services.