Network.AWS.Support.V2013_04_15.DescribeSeverityLevels

Description

Returns the list of severity levels that you can assign to an AWS Support case. The severity level for a case is also a field in the CaseDetails data type included in any CreateCase request.

Request

data DescribeSeverityLevels

Request constructor

mkDescribeSeverityLevels :: DescribeSeverityLevels

Smart constructor for the minimum required parameters to construct a valid DescribeSeverityLevels request.

Request lenses

dslLanguage :: Lens' DescribeSeverityLevels (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Response

data DescribeSeverityLevelsResponse

The list of severity levels returned by the DescribeSeverityLevels operation.

Response lenses

dslrSeverityLevels :: Lens' DescribeSeverityLevelsResponse [SeverityLevel]

The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.