Network.AWS.Support.V2013_04_15.DescribeTrustedAdvisorCheckResult

Description

Returns the results of the Trusted Advisor check that has the specified check ID. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks. The response contains a TrustedAdvisorCheckResult object, which contains these three objects: TrustedAdvisorCategorySpecificSummary TrustedAdvisorResourceDetail TrustedAdvisorResourcesSummary In addition, the response contains these fields: Status. The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available". Timestamp. The time of the last refresh of the check. CheckId. The unique identifier for the check.

Documentation

describeTrustedAdvisorCheckResult

Arguments

:: Text
_dtacrrCheckId

-> DescribeTrustedAdvisorCheckResult

Minimum specification for a DescribeTrustedAdvisorCheckResult request.

data DescribeTrustedAdvisorCheckResult

Constructors

DescribeTrustedAdvisorCheckResult

Fields

_dtacrrCheckId :: Text

The unique identifier for the Trusted Advisor check.

_dtacrrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

dtacrrLanguage :: Lens' DescribeTrustedAdvisorCheckResult (Maybe Text)

dtacrrCheckId :: Lens' DescribeTrustedAdvisorCheckResult Text

data DescribeTrustedAdvisorCheckResultResponse

Constructors

DescribeTrustedAdvisorCheckResultResponse

Fields

_dtacrsResult :: Maybe TrustedAdvisorCheckResult

The detailed results of the Trusted Advisor check.

dtacrsResult :: Iso' DescribeTrustedAdvisorCheckResultResponse (Maybe TrustedAdvisorCheckResult)