Network.AWS.Support.V2013_04_15.DescribeTrustedAdvisorChecks

Description

Returns information about all available Trusted Advisor checks, including name, ID, category, description, and metadata. You must specify a language code; English ("en") and Japanese ("ja") are currently supported. The response contains a TrustedAdvisorCheckDescription for each check.

Request

data DescribeTrustedAdvisorChecks

Request constructor

mkDescribeTrustedAdvisorChecks

Arguments

:: Text
dtacLanguage

-> DescribeTrustedAdvisorChecks

Smart constructor for the minimum required parameters to construct a valid DescribeTrustedAdvisorChecks request.

Request lenses

dtacLanguage :: Lens' DescribeTrustedAdvisorChecks Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Response

data DescribeTrustedAdvisorChecksResponse

Information about the Trusted Advisor checks returned by the DescribeTrustedAdvisorChecks operation.

Response lenses

dtacr1Checks :: Lens' DescribeTrustedAdvisorChecksResponse [TrustedAdvisorCheckDescription]

Information about all available Trusted Advisor checks.