Network.AWS.Support.V2013_04_15.RefreshTrustedAdvisorCheck

Description

Requests a refresh of the Trusted Advisor check that has the specified check ID. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks. The response contains a RefreshTrustedAdvisorCheckResult object, which contains these fields: Status. The refresh status of the check: "none", "enqueued", "processing", "success", or "abandoned". MillisUntilNextRefreshable. The amount of time, in milliseconds, until the check is eligible for refresh. CheckId. The unique identifier for the check.

Documentation

data RefreshTrustedAdvisorCheck

Constructors

RefreshTrustedAdvisorCheck

Fields

_rtacrCheckId :: Text

The unique identifier for the Trusted Advisor check.

rtacrCheckId :: Iso' RefreshTrustedAdvisorCheck Text

data RefreshTrustedAdvisorCheckResponse

Constructors

RefreshTrustedAdvisorCheckResponse

Fields

_rtacsStatus :: TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.

rtacsStatus :: Iso' RefreshTrustedAdvisorCheckResponse TrustedAdvisorCheckRefreshStatus