Network.AWS.Support.V2013_04_15.ResolveCase

Description

Takes a CaseId and returns the initial state of the case along with the state of the case after the call to ResolveCase completed.

Documentation

resolveCase :: ResolveCase

Minimum specification for a ResolveCase request.

data ResolveCase

Constructors

ResolveCase

Fields

_rcrCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

rcrCaseId :: Iso' ResolveCase (Maybe Text)

data ResolveCaseResponse

Constructors

ResolveCaseResponse

Fields

_rcsInitialCaseStatus :: Maybe Text

The status of the case when the ResolveCase request was sent.

_rcsFinalCaseStatus :: Maybe Text

The status of the case after the ResolveCase request was processed.

rcsInitialCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)

rcsFinalCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)