Network.AWS.Support.V2013_04_15.Types

Description

AWS Support is a one-on-one, fast-response support channel that is staffed with experienced support personnel. The service helps customers get the most from the products and features provided by Amazon Web Services. There are four levels, or tiers, of AWS Support: Basic, Developer, Business, and Enterprise. The Basic tier is free of charge and offers support for account and billing questions and service limit increases. The other tiers offer an unlimited number of technical support cases with pay-by-the-month pricing and no long-term contracts, providing developers and businesses flexibility to choose the level of support that meets their needs.

Documentation

data Support

Supported version (2013-04-15) of the AWS Support service.

newtype TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

Constructors

TrustedAdvisorCategorySpecificSummary

Fields

_tacssCostOptimizing :: Maybe TrustedAdvisorCostOptimizingSummary

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

data Attachment

The attachment content and file name.

Constructors

Attachment

Fields

_aData :: Maybe Base64

The content of the attachment file.

_aFileName :: Maybe Text

The name of the attachment file.

data AttachmentDetails

The file name and ID of an attachment to a case communication. You can use the ID to retrieve the attachment with the DescribeAttachment operation.

Constructors

AttachmentDetails

Fields

_adAttachmentId :: Maybe Text

The ID of the attachment.

_adFileName :: Maybe Text

The file name of the attachment.

data CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained the response from a DescribeCases request. CaseDetails contains the following fields: CaseID. The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47. CategoryCode. The category of problem for the AWS Support case. Corresponds to the CategoryCode values returned by a call to DescribeServices. DisplayId. The identifier for the case on pages in the AWS Support Center. Language. The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them. RecentCommunications. One or more Communication objects. Fields of these objects are Attachments, Body, CaseId, SubmittedBy, and TimeCreated. NextToken. A resumption point for pagination. ServiceCode. The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices. SeverityCode. The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels. Status. The status of the case in the AWS Support Center. Subject. The subject line of the case. SubmittedBy. The email address of the account that submitted the case. TimeCreated. The time the case was created, in ISO-8601 format.

Constructors

CaseDetails

Fields

_cdSubject :: Maybe Text

The subject line for the case in the AWS Support Center.

_cdStatus :: Maybe Text

The status of the case.

_cdRecentCommunications :: Maybe RecentCaseCommunications

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

_cdSeverityCode :: Maybe Text

The code for the severity level returned by the call to DescribeSeverityLevels.

_cdCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_cdCcEmailAddresses :: [Text]

The email addresses that receive copies of communication about the case.

_cdDisplayId :: Maybe Text

The ID displayed for the case in the AWS Support Center. This is a numeric string.

_cdSubmittedBy :: Maybe Text

The email address of the account that submitted the case.

_cdLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

_cdTimeCreated :: Maybe Text

The time that the case was case created in the AWS Support Center.

_cdCategoryCode :: Maybe Text

The category of problem for the AWS Support case.

_cdServiceCode :: Maybe Text

The code for the AWS service returned by the call to DescribeServices.

data Category

A JSON-formatted name/value pair that represents the category name and category code of the problem, selected from the DescribeServices response for each AWS service.

Constructors

Category

Fields

_cyName :: Maybe Text

The category name for the support case.

_cyCode :: Maybe Text

The category code for the support case.

data Communication

A communication associated with an AWS Support case. The communication consists of the case ID, the message body, attachment information, the account email address, and the date and time of the communication.

Constructors

Communication

Fields

_cBody :: Maybe Text

The text of the communication between the customer and AWS Support.

_cCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_cSubmittedBy :: Maybe Text

The email address of the account that submitted the AWS Support case.

_cTimeCreated :: Maybe Text

The time the communication was created.

_cAttachmentSet :: [AttachmentDetails]

Information about the attachments to the case communication.

data RecentCaseCommunications

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

Constructors

RecentCaseCommunications

Fields

_rccNextToken :: Maybe Text

A resumption point for pagination.

_rccCommunications :: [Communication]

The five most recent communications associated with the case.

data Service

Information about an AWS service returned by the DescribeServices operation.

Constructors

Service

Fields

_ssfCategories :: [Category]

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

_ssfName :: Maybe Text

The friendly name for an AWS service. The Code element contains the corresponding code.

_ssfCode :: Maybe Text

The code for an AWS service returned by the DescribeServices response. The Name element contains the corresponding friendly name.

data SeverityLevel

A code and name pair that represent a severity level that can be applied to a support case.

Constructors

SeverityLevel

Fields

_snName :: Maybe Text

The name of the severity level that corresponds to the severity level code.

_snCode :: Maybe Text

One of four values: "low," "medium," "high," and "urgent". These values correspond to response times returned to the caller in SeverityLevel.name.

data TrustedAdvisorCheckDescription

The description and metadata for a Trusted Advisor check.

Constructors

TrustedAdvisorCheckDescription

Fields

_tacdCategory :: Text

The category of the Trusted Advisor check.

_tacdName :: Text

The display name for the Trusted Advisor check.

_tacdMetadata :: [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

_tacdId :: Text

The unique identifier for the Trusted Advisor check.

_tacdDescription :: Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended actions (contains HTML markup).

data TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.

Constructors

TrustedAdvisorCheckRefreshStatus

Fields

_tacrsMillisUntilNextRefreshable :: Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

_tacrsStatus :: Text

The status of the Trusted Advisor check for which a refresh has been requested: "none", "enqueued", "processing", "success", or "abandoned".

_tacrsCheckId :: Text

The unique identifier for the Trusted Advisor check.

data TrustedAdvisorCheckResult

The detailed results of the Trusted Advisor check.

Constructors

TrustedAdvisorCheckResult

Fields

_tacrCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

_tacrStatus :: Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

_tacrCheckId :: Text

The unique identifier for the Trusted Advisor check.

_tacrFlaggedResources :: [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

_tacrResourcesSummary :: TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

_tacrTimestamp :: Text

The time of the last refresh of the check.

data TrustedAdvisorCheckSummary

A summary of a Trusted Advisor check result, including the alert status, last refresh, and number of resources examined.

Constructors

TrustedAdvisorCheckSummary

Fields

_tacsCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

_tacsStatus :: Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

_tacsCheckId :: Text

The unique identifier for the Trusted Advisor check.

_tacsResourcesSummary :: TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

_tacsTimestamp :: Text

The time of the last refresh of the check.

_tacsHasFlaggedResources :: Maybe Bool

Specifies whether the Trusted Advisor check has flagged resources.

data TrustedAdvisorCostOptimizingSummary

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

Constructors

TrustedAdvisorCostOptimizingSummary

Fields

_tacosEstimatedMonthlySavings :: Double

The estimated monthly savings that might be realized if the recommended actions are taken.

_tacosEstimatedPercentMonthlySavings :: Double

The estimated percentage of savings that might be realized if the recommended actions are taken.

data TrustedAdvisorResourceDetail

Contains information about a resource identified by a Trusted Advisor check.

Constructors

TrustedAdvisorResourceDetail

Fields

_tardStatus :: Text

The status code for the resource identified in the Trusted Advisor check.

_tardResourceId :: Text

The unique identifier for the identified resource.

_tardIsSuppressed :: Maybe Bool

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

_tardMetadata :: [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

_tardRegion :: Text

The AWS region in which the identified resource is located.

data TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

Constructors

TrustedAdvisorResourcesSummary

Fields

_tarsResourcesIgnored :: Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

_tarsResourcesProcessed :: Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

_tarsResourcesFlagged :: Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

_tarsResourcesSuppressed :: Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

tacssCostOptimizing :: Iso' TrustedAdvisorCategorySpecificSummary (Maybe TrustedAdvisorCostOptimizingSummary)

aFileName :: Lens' Attachment (Maybe Text)

aData :: Lens' Attachment (Maybe Base64)

adFileName :: Lens' AttachmentDetails (Maybe Text)

adAttachmentId :: Lens' AttachmentDetails (Maybe Text)

cdTimeCreated :: Lens' CaseDetails (Maybe Text)

cdSubmittedBy :: Lens' CaseDetails (Maybe Text)

cdSubject :: Lens' CaseDetails (Maybe Text)

cdStatus :: Lens' CaseDetails (Maybe Text)

cdSeverityCode :: Lens' CaseDetails (Maybe Text)

cdServiceCode :: Lens' CaseDetails (Maybe Text)

cdRecentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

cdLanguage :: Lens' CaseDetails (Maybe Text)

cdDisplayId :: Lens' CaseDetails (Maybe Text)

cdCcEmailAddresses :: Lens' CaseDetails [Text]

cdCategoryCode :: Lens' CaseDetails (Maybe Text)

cdCaseId :: Lens' CaseDetails (Maybe Text)

cyName :: Lens' Category (Maybe Text)

cyCode :: Lens' Category (Maybe Text)

cTimeCreated :: Lens' Communication (Maybe Text)

cSubmittedBy :: Lens' Communication (Maybe Text)

cCaseId :: Lens' Communication (Maybe Text)

cBody :: Lens' Communication (Maybe Text)

cAttachmentSet :: Lens' Communication [AttachmentDetails]

rccNextToken :: Lens' RecentCaseCommunications (Maybe Text)

rccCommunications :: Lens' RecentCaseCommunications [Communication]

ssfName :: Lens' Service (Maybe Text)

ssfCode :: Lens' Service (Maybe Text)

ssfCategories :: Lens' Service [Category]

snName :: Lens' SeverityLevel (Maybe Text)

snCode :: Lens' SeverityLevel (Maybe Text)

tacdName :: Lens' TrustedAdvisorCheckDescription Text

tacdMetadata :: Lens' TrustedAdvisorCheckDescription [Text]

tacdId :: Lens' TrustedAdvisorCheckDescription Text

tacdDescription :: Lens' TrustedAdvisorCheckDescription Text

tacdCategory :: Lens' TrustedAdvisorCheckDescription Text

tacrsStatus :: Lens' TrustedAdvisorCheckRefreshStatus Text

tacrsMillisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

tacrsCheckId :: Lens' TrustedAdvisorCheckRefreshStatus Text

tacrTimestamp :: Lens' TrustedAdvisorCheckResult Text

tacrStatus :: Lens' TrustedAdvisorCheckResult Text

tacrResourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

tacrFlaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

tacrCheckId :: Lens' TrustedAdvisorCheckResult Text

tacrCategorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

tacsTimestamp :: Lens' TrustedAdvisorCheckSummary Text

tacsStatus :: Lens' TrustedAdvisorCheckSummary Text

tacsResourcesSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorResourcesSummary

tacsHasFlaggedResources :: Lens' TrustedAdvisorCheckSummary (Maybe Bool)

tacsCheckId :: Lens' TrustedAdvisorCheckSummary Text

tacsCategorySpecificSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorCategorySpecificSummary

tacosEstimatedPercentMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

tacosEstimatedMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

tardStatus :: Lens' TrustedAdvisorResourceDetail Text

tardResourceId :: Lens' TrustedAdvisorResourceDetail Text

tardRegion :: Lens' TrustedAdvisorResourceDetail Text

tardMetadata :: Lens' TrustedAdvisorResourceDetail [Text]

tardIsSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

tarsResourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

tarsResourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

tarsResourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

tarsResourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer