Network.AWS.Support

Description

AWS Support

The AWS Support API reference is intended for programmers who need detailed information about the AWS Support operations and data types. This service enables you to manage your AWS Support cases programmatically. It uses HTTP methods that return results in JSON format.

The AWS Support service also exposes a set of Trusted Advisor features. You can retrieve a list of checks and their descriptions, get check results, specify checks to refresh, and get the refresh status of checks.

The following list describes the AWS Support case management operations:

Service names, issue categories, and available severity levels. The DescribeServices and DescribeSeverityLevels operations return AWS service names, service codes, service categories, and problem severity levels. You use these values when you call the CreateCase operation.

Case creation, case details, and case resolution. The CreateCase , DescribeCases , DescribeAttachment , and ResolveCase operations create AWS Support cases, retrieve information about cases, and resolve cases.

Case communication. The DescribeCommunications , AddCommunicationToCase , and AddAttachmentsToSet operations retrieve and add communications and attachments to AWS Support cases.

The following list describes the operations available from the AWS Support service for Trusted Advisor:

DescribeTrustedAdvisorChecks returns the list of checks that run against your AWS resources.

Using the checkId for a specific check returned by DescribeTrustedAdvisorChecks , you can call DescribeTrustedAdvisorCheckResult to obtain the results for the check you specified.

DescribeTrustedAdvisorCheckSummaries returns summarized results for one or more Trusted Advisor checks.

RefreshTrustedAdvisorCheck requests that Trusted Advisor rerun a specified check.

DescribeTrustedAdvisorCheckRefreshStatuses reports the refresh status of one or more checks.

For authentication of requests, AWS Support uses Signature Version 4 Signing Process .

See About the AWS Support API in the AWS Support User Guide for information about how to use this service to create and manage your support cases, and how to call Trusted Advisor for results of checks on your resources.

Service Configuration

support :: Service

API version 2013-04-15 of the Amazon Support SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Support.

AttachmentSetExpired

_AttachmentSetExpired :: AsError a => Getting (First ServiceError) a ServiceError

The expiration time of the attachment set has passed. The set expires 1 hour after it is created.

AttachmentLimitExceeded

_AttachmentLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The limit for the number of attachment sets created in a short period of time has been exceeded.

DescribeAttachmentLimitExceeded

_DescribeAttachmentLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The limit for the number of DescribeAttachment requests in a short period of time has been exceeded.

CaseIdNotFound

_CaseIdNotFound :: AsError a => Getting (First ServiceError) a ServiceError

The requested caseId could not be located.

AttachmentSetIdNotFound

_AttachmentSetIdNotFound :: AsError a => Getting (First ServiceError) a ServiceError

An attachment set with the specified ID could not be found.

AttachmentSetSizeLimitExceeded

_AttachmentSetSizeLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

A limit for the size of an attachment set has been exceeded. The limits are 3 attachments and 5 MB per attachment.

AttachmentIdNotFound

_AttachmentIdNotFound :: AsError a => Getting (First ServiceError) a ServiceError

An attachment with the specified ID could not be found.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred.

CaseCreationLimitExceeded

_CaseCreationLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The case creation limit for the account has been exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

RefreshTrustedAdvisorCheck

module Network.AWS.Support.RefreshTrustedAdvisorCheck

DescribeCases (Paginated)

module Network.AWS.Support.DescribeCases

DescribeTrustedAdvisorCheckRefreshStatuses

module Network.AWS.Support.DescribeTrustedAdvisorCheckRefreshStatuses

DescribeTrustedAdvisorCheckSummaries

module Network.AWS.Support.DescribeTrustedAdvisorCheckSummaries

CreateCase

module Network.AWS.Support.CreateCase

ResolveCase

module Network.AWS.Support.ResolveCase

DescribeSeverityLevels

module Network.AWS.Support.DescribeSeverityLevels

DescribeTrustedAdvisorChecks

module Network.AWS.Support.DescribeTrustedAdvisorChecks

DescribeAttachment

module Network.AWS.Support.DescribeAttachment

AddAttachmentsToSet

module Network.AWS.Support.AddAttachmentsToSet

DescribeTrustedAdvisorCheckResult

module Network.AWS.Support.DescribeTrustedAdvisorCheckResult

DescribeServices

module Network.AWS.Support.DescribeServices

DescribeCommunications (Paginated)

module Network.AWS.Support.DescribeCommunications

AddCommunicationToCase

module Network.AWS.Support.AddCommunicationToCase

Types

Attachment

data Attachment

An attachment to a case communication. The attachment consists of the file name and the content of the file.

See: attachment smart constructor.

attachment :: Attachment

Creates a value of Attachment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aData - The content of the attachment file.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

aFileName - The name of the attachment file.

aData :: Lens' Attachment (Maybe ByteString)

The content of the attachment file.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

aFileName :: Lens' Attachment (Maybe Text)

The name of the attachment file.

AttachmentDetails

data AttachmentDetails

The file name and ID of an attachment to a case communication. You can use the ID to retrieve the attachment with the DescribeAttachment operation.

See: attachmentDetails smart constructor.

attachmentDetails :: AttachmentDetails

Creates a value of AttachmentDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adAttachmentId - The ID of the attachment.

adFileName - The file name of the attachment.

adAttachmentId :: Lens' AttachmentDetails (Maybe Text)

The ID of the attachment.

adFileName :: Lens' AttachmentDetails (Maybe Text)

The file name of the attachment.

CaseDetails

data CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained the response from a DescribeCases request. CaseDetails contains the following fields:

caseId. The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47 .

categoryCode. The category of problem for the AWS Support case. Corresponds to the CategoryCode values returned by a call to DescribeServices .

displayId. The identifier for the case on pages in the AWS Support Center.

language. The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

recentCommunications. One or more Communication objects. Fields of these objects are attachments , body , caseId , submittedBy , and timeCreated .

nextToken. A resumption point for pagination.

serviceCode. The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices .

severityCode. The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels .

status. The status of the case in the AWS Support Center.

subject. The subject line of the case.

submittedBy. The email address of the account that submitted the case.

timeCreated. The time the case was created, in ISO-8601 format.

See: caseDetails smart constructor.

caseDetails :: CaseDetails

Creates a value of CaseDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdSubject - The subject line for the case in the AWS Support Center.

cdStatus - The status of the case.

cdRecentCommunications - The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

cdSeverityCode - The code for the severity level returned by the call to DescribeSeverityLevels .

cdCaseId - The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

cdCcEmailAddresses - The email addresses that receive copies of communication about the case.

cdDisplayId - The ID displayed for the case in the AWS Support Center. This is a numeric string.

cdSubmittedBy - The email address of the account that submitted the case.

cdLanguage - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

cdTimeCreated - The time that the case was case created in the AWS Support Center.

cdCategoryCode - The category of problem for the AWS Support case.

cdServiceCode - The code for the AWS service returned by the call to DescribeServices .

cdSubject :: Lens' CaseDetails (Maybe Text)

The subject line for the case in the AWS Support Center.

cdStatus :: Lens' CaseDetails (Maybe Text)

The status of the case.

cdRecentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

cdSeverityCode :: Lens' CaseDetails (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels .

cdCaseId :: Lens' CaseDetails (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

cdCcEmailAddresses :: Lens' CaseDetails [Text]

The email addresses that receive copies of communication about the case.

cdDisplayId :: Lens' CaseDetails (Maybe Text)

The ID displayed for the case in the AWS Support Center. This is a numeric string.

cdSubmittedBy :: Lens' CaseDetails (Maybe Text)

The email address of the account that submitted the case.

cdLanguage :: Lens' CaseDetails (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

cdTimeCreated :: Lens' CaseDetails (Maybe Text)

The time that the case was case created in the AWS Support Center.

cdCategoryCode :: Lens' CaseDetails (Maybe Text)

The category of problem for the AWS Support case.

cdServiceCode :: Lens' CaseDetails (Maybe Text)

The code for the AWS service returned by the call to DescribeServices .

Category

data Category

A JSON-formatted name/value pair that represents the category name and category code of the problem, selected from the DescribeServices response for each AWS service.

See: category smart constructor.

category :: Category

Creates a value of Category with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cName - The category name for the support case.

cCode - The category code for the support case.

cName :: Lens' Category (Maybe Text)

The category name for the support case.

cCode :: Lens' Category (Maybe Text)

The category code for the support case.

Communication

data Communication

A communication associated with an AWS Support case. The communication consists of the case ID, the message body, attachment information, the account email address, and the date and time of the communication.

See: communication smart constructor.

communication :: Communication

Creates a value of Communication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cBody - The text of the communication between the customer and AWS Support.

cCaseId - The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

cSubmittedBy - The email address of the account that submitted the AWS Support case.

cTimeCreated - The time the communication was created.

cAttachmentSet - Information about the attachments to the case communication.

cBody :: Lens' Communication (Maybe Text)

The text of the communication between the customer and AWS Support.

cCaseId :: Lens' Communication (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

cSubmittedBy :: Lens' Communication (Maybe Text)

The email address of the account that submitted the AWS Support case.

cTimeCreated :: Lens' Communication (Maybe Text)

The time the communication was created.

cAttachmentSet :: Lens' Communication [AttachmentDetails]

Information about the attachments to the case communication.

RecentCaseCommunications

data RecentCaseCommunications

The five most recent communications associated with the case.

See: recentCaseCommunications smart constructor.

recentCaseCommunications :: RecentCaseCommunications

Creates a value of RecentCaseCommunications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rccNextToken - A resumption point for pagination.

rccCommunications - The five most recent communications associated with the case.

rccNextToken :: Lens' RecentCaseCommunications (Maybe Text)

A resumption point for pagination.

rccCommunications :: Lens' RecentCaseCommunications [Communication]

The five most recent communications associated with the case.

SeverityLevel

data SeverityLevel

A code and name pair that represent a severity level that can be applied to a support case.

See: severityLevel smart constructor.

severityLevel :: SeverityLevel

Creates a value of SeverityLevel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slName - The name of the severity level that corresponds to the severity level code.

slCode - One of four values: "low," "medium," "high," and "urgent". These values correspond to response times returned to the caller in severityLevel.name .

slName :: Lens' SeverityLevel (Maybe Text)

The name of the severity level that corresponds to the severity level code.

slCode :: Lens' SeverityLevel (Maybe Text)

One of four values: "low," "medium," "high," and "urgent". These values correspond to response times returned to the caller in severityLevel.name .

SupportService

data SupportService

Information about an AWS service returned by the DescribeServices operation.

See: supportService smart constructor.

supportService :: SupportService

Creates a value of SupportService with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssCategories - A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase .

ssName - The friendly name for an AWS service. The code element contains the corresponding code.

ssCode - The code for an AWS service returned by the DescribeServices response. The name element contains the corresponding friendly name.

ssCategories :: Lens' SupportService [Category]

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase .

ssName :: Lens' SupportService (Maybe Text)

The friendly name for an AWS service. The code element contains the corresponding code.

ssCode :: Lens' SupportService (Maybe Text)

The code for an AWS service returned by the DescribeServices response. The name element contains the corresponding friendly name.

TrustedAdvisorCategorySpecificSummary

data TrustedAdvisorCategorySpecificSummary

The container for summary information that relates to the category of the Trusted Advisor check.

See: trustedAdvisorCategorySpecificSummary smart constructor.

trustedAdvisorCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Creates a value of TrustedAdvisorCategorySpecificSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tacssCostOptimizing - The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

tacssCostOptimizing :: Lens' TrustedAdvisorCategorySpecificSummary (Maybe TrustedAdvisorCostOptimizingSummary)

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

TrustedAdvisorCheckDescription

data TrustedAdvisorCheckDescription

The description and metadata for a Trusted Advisor check.

See: trustedAdvisorCheckDescription smart constructor.

trustedAdvisorCheckDescription

Arguments

:: Text
tacdId

-> Text
tacdName

-> Text
tacdDescription

-> Text
tacdCategory

-> TrustedAdvisorCheckDescription

Creates a value of TrustedAdvisorCheckDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tacdId - The unique identifier for the Trusted Advisor check.

tacdName - The display name for the Trusted Advisor check.

tacdDescription - The description of the Trusted Advisor check, which includes the alert criteria and recommended actions (contains HTML markup).

tacdCategory - The category of the Trusted Advisor check.

tacdMetadata - The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

tacdId :: Lens' TrustedAdvisorCheckDescription Text

The unique identifier for the Trusted Advisor check.

tacdName :: Lens' TrustedAdvisorCheckDescription Text

The display name for the Trusted Advisor check.

tacdDescription :: Lens' TrustedAdvisorCheckDescription Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended actions (contains HTML markup).

tacdCategory :: Lens' TrustedAdvisorCheckDescription Text

The category of the Trusted Advisor check.

tacdMetadata :: Lens' TrustedAdvisorCheckDescription [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorCheckRefreshStatus

data TrustedAdvisorCheckRefreshStatus

The refresh status of a Trusted Advisor check.

See: trustedAdvisorCheckRefreshStatus smart constructor.

trustedAdvisorCheckRefreshStatus

Arguments

:: Text
tacrsCheckId

-> Text
tacrsStatus

-> Integer
tacrsMillisUntilNextRefreshable

-> TrustedAdvisorCheckRefreshStatus

Creates a value of TrustedAdvisorCheckRefreshStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tacrsCheckId - The unique identifier for the Trusted Advisor check.

tacrsStatus - The status of the Trusted Advisor check for which a refresh has been requested: "none", "enqueued", "processing", "success", or "abandoned".

tacrsMillisUntilNextRefreshable - The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

tacrsCheckId :: Lens' TrustedAdvisorCheckRefreshStatus Text

The unique identifier for the Trusted Advisor check.

tacrsStatus :: Lens' TrustedAdvisorCheckRefreshStatus Text

The status of the Trusted Advisor check for which a refresh has been requested: "none", "enqueued", "processing", "success", or "abandoned".

tacrsMillisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

TrustedAdvisorCheckResult

data TrustedAdvisorCheckResult

The results of a Trusted Advisor check returned by DescribeTrustedAdvisorCheckResult .

See: trustedAdvisorCheckResult smart constructor.

trustedAdvisorCheckResult

Arguments

:: Text
tacrCheckId

-> Text
tacrTimestamp

-> Text
tacrStatus

-> TrustedAdvisorResourcesSummary
tacrResourcesSummary

-> TrustedAdvisorCategorySpecificSummary
tacrCategorySpecificSummary

-> TrustedAdvisorCheckResult

Creates a value of TrustedAdvisorCheckResult with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tacrCheckId - The unique identifier for the Trusted Advisor check.

tacrTimestamp - The time of the last refresh of the check.

tacrStatus - The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacrResourcesSummary - Undocumented member.

tacrCategorySpecificSummary - Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

tacrFlaggedResources - The details about each resource listed in the check result.

tacrCheckId :: Lens' TrustedAdvisorCheckResult Text

The unique identifier for the Trusted Advisor check.

tacrTimestamp :: Lens' TrustedAdvisorCheckResult Text

The time of the last refresh of the check.

tacrStatus :: Lens' TrustedAdvisorCheckResult Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacrResourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

Undocumented member.

tacrCategorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

tacrFlaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

TrustedAdvisorCheckSummary

data TrustedAdvisorCheckSummary

A summary of a Trusted Advisor check result, including the alert status, last refresh, and number of resources examined.

See: trustedAdvisorCheckSummary smart constructor.

trustedAdvisorCheckSummary

Arguments

:: Text
tacsCheckId

-> Text
tacsTimestamp

-> Text
tacsStatus

-> TrustedAdvisorResourcesSummary
tacsResourcesSummary

-> TrustedAdvisorCategorySpecificSummary
tacsCategorySpecificSummary

-> TrustedAdvisorCheckSummary

Creates a value of TrustedAdvisorCheckSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tacsHasFlaggedResources - Specifies whether the Trusted Advisor check has flagged resources.

tacsCheckId - The unique identifier for the Trusted Advisor check.

tacsTimestamp - The time of the last refresh of the check.

tacsStatus - The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacsResourcesSummary - Undocumented member.

tacsCategorySpecificSummary - Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

tacsHasFlaggedResources :: Lens' TrustedAdvisorCheckSummary (Maybe Bool)

Specifies whether the Trusted Advisor check has flagged resources.

tacsCheckId :: Lens' TrustedAdvisorCheckSummary Text

The unique identifier for the Trusted Advisor check.

tacsTimestamp :: Lens' TrustedAdvisorCheckSummary Text

The time of the last refresh of the check.

tacsStatus :: Lens' TrustedAdvisorCheckSummary Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacsResourcesSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorResourcesSummary

Undocumented member.

tacsCategorySpecificSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

TrustedAdvisorCostOptimizingSummary

data TrustedAdvisorCostOptimizingSummary

The estimated cost savings that might be realized if the recommended actions are taken.

See: trustedAdvisorCostOptimizingSummary smart constructor.

trustedAdvisorCostOptimizingSummary

Arguments

:: Double
tacosEstimatedMonthlySavings

-> Double
tacosEstimatedPercentMonthlySavings

-> TrustedAdvisorCostOptimizingSummary

Creates a value of TrustedAdvisorCostOptimizingSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tacosEstimatedMonthlySavings - The estimated monthly savings that might be realized if the recommended actions are taken.

tacosEstimatedPercentMonthlySavings - The estimated percentage of savings that might be realized if the recommended actions are taken.

tacosEstimatedMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated monthly savings that might be realized if the recommended actions are taken.

tacosEstimatedPercentMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated percentage of savings that might be realized if the recommended actions are taken.

TrustedAdvisorResourceDetail

data TrustedAdvisorResourceDetail

Contains information about a resource identified by a Trusted Advisor check.

See: trustedAdvisorResourceDetail smart constructor.

trustedAdvisorResourceDetail

Arguments

:: Text
tardStatus

-> Text
tardResourceId

-> TrustedAdvisorResourceDetail

Creates a value of TrustedAdvisorResourceDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tardIsSuppressed - Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

tardRegion - The AWS region in which the identified resource is located.

tardStatus - The status code for the resource identified in the Trusted Advisor check.

tardResourceId - The unique identifier for the identified resource.

tardMetadata - Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks . Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

tardIsSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

tardRegion :: Lens' TrustedAdvisorResourceDetail (Maybe Text)

The AWS region in which the identified resource is located.

tardStatus :: Lens' TrustedAdvisorResourceDetail Text

The status code for the resource identified in the Trusted Advisor check.

tardResourceId :: Lens' TrustedAdvisorResourceDetail Text

The unique identifier for the identified resource.

tardMetadata :: Lens' TrustedAdvisorResourceDetail [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks . Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorResourcesSummary

data TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries .

See: trustedAdvisorResourcesSummary smart constructor.

trustedAdvisorResourcesSummary

Arguments

:: Integer
tarsResourcesProcessed

-> Integer
tarsResourcesFlagged

-> Integer
tarsResourcesIgnored

-> Integer
tarsResourcesSuppressed

-> TrustedAdvisorResourcesSummary

Creates a value of TrustedAdvisorResourcesSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tarsResourcesProcessed - The number of AWS resources that were analyzed by the Trusted Advisor check.

tarsResourcesFlagged - The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

tarsResourcesIgnored - The number of AWS resources ignored by Trusted Advisor because information was unavailable.

tarsResourcesSuppressed - The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

tarsResourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

tarsResourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

tarsResourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

tarsResourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.