Network.AWS.Support.AddCommunicationToCase

Description

Adds additional customer communication to an AWS Support case. You use the caseId value to identify the case to add communication to. You can list a set of email addresses to copy on the communication using the ccEmailAddresses value. The communicationBody value contains the text of the communication.

The response indicates the success or failure of the request.

This operation implements a subset of the features of the AWS Support Center.

Creating a Request

addCommunicationToCase

Arguments

:: Text
actcCommunicationBody

-> AddCommunicationToCase

Creates a value of AddCommunicationToCase with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

actcCaseId - The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

actcCcEmailAddresses - The email addresses in the CC line of an email to be added to the support case.

actcAttachmentSetId - The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet

actcCommunicationBody - The body of an email communication to add to the support case.

data AddCommunicationToCase

To be written.

See: addCommunicationToCase smart constructor.

Request Lenses

actcCaseId :: Lens' AddCommunicationToCase (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

actcCcEmailAddresses :: Lens' AddCommunicationToCase [Text]

The email addresses in the CC line of an email to be added to the support case.

actcAttachmentSetId :: Lens' AddCommunicationToCase (Maybe Text)

The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet

actcCommunicationBody :: Lens' AddCommunicationToCase Text

The body of an email communication to add to the support case.

Destructuring the Response

addCommunicationToCaseResponse

Arguments

:: Int
actcrsResponseStatus

-> AddCommunicationToCaseResponse

Creates a value of AddCommunicationToCaseResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

actcrsResult - True if AddCommunicationToCase succeeds. Otherwise, returns an error.

actcrsResponseStatus - -- | The response status code.

data AddCommunicationToCaseResponse

The result of the AddCommunicationToCase operation.

See: addCommunicationToCaseResponse smart constructor.

Response Lenses

actcrsResult :: Lens' AddCommunicationToCaseResponse (Maybe Bool)

True if AddCommunicationToCase succeeds. Otherwise, returns an error.

actcrsResponseStatus :: Lens' AddCommunicationToCaseResponse Int

- | The response status code.