Network.AWS.Support.AddCommunicationToCase

Description

Adds additional customer communication to an AWS Support case. You use the CaseId value to identify the case to add communication to. You can list a set of email addresses to copy on the communication using the CcEmailAddresses value. The CommunicationBody value contains the text of the communication.

The response indicates the success or failure of the request.

This operation implements a subset of the features of the AWS Support Center.

http://docs.aws.amazon.com/awssupport/latest/APIReference/API_AddCommunicationToCase.html

Request

data AddCommunicationToCase

Request constructor

addCommunicationToCase

Arguments

:: Text
actcCommunicationBody

-> AddCommunicationToCase

AddCommunicationToCase constructor.

The fields accessible through corresponding lenses are:

actcAttachmentSetId :: Maybe Text

actcCaseId :: Maybe Text

actcCcEmailAddresses :: [Text]

actcCommunicationBody :: Text

Request lenses

actcAttachmentSetId :: Lens' AddCommunicationToCase (Maybe Text)

The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet

actcCaseId :: Lens' AddCommunicationToCase (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

actcCcEmailAddresses :: Lens' AddCommunicationToCase [Text]

The email addresses in the CC line of an email to be added to the support case.

actcCommunicationBody :: Lens' AddCommunicationToCase Text

The body of an email communication to add to the support case.

Response

data AddCommunicationToCaseResponse

Response constructor

addCommunicationToCaseResponse :: AddCommunicationToCaseResponse

AddCommunicationToCaseResponse constructor.

The fields accessible through corresponding lenses are:

actcrResult :: Maybe Bool

Response lenses

actcrResult :: Lens' AddCommunicationToCaseResponse (Maybe Bool)

True if AddCommunicationToCase succeeds. Otherwise, returns an error.