Network.AWS.Support.CreateCase

Description

Creates a new case in the AWS Support Center. This operation is modeled on the behavior of the AWS Support Center Create Case page. Its parameters require you to specify the following information:

issueType. The type of issue for the case. You can specify either "customer-service" or "technical." If you do not indicate a value, the default is "technical."

serviceCode. The code for an AWS service. You obtain the serviceCode by calling DescribeServices .

categoryCode. The category for the service defined for the serviceCode value. You also obtain the category code for a service by calling DescribeServices . Each AWS service defines its own set of category codes.

severityCode. A value that indicates the urgency of the case, which in turn determines the response time according to your service level agreement with AWS Support. You obtain the SeverityCode by calling DescribeSeverityLevels .

subject. The Subject field on the AWS Support Center Create Case page.

communicationBody. The Description field on the AWS Support Center Create Case page.

attachmentSetId. The ID of a set of attachments that has been created by using AddAttachmentsToSet .

language. The human language in which AWS Support handles the case. English and Japanese are currently supported.

ccEmailAddresses. The AWS Support Center CC field on the Create Case page. You can list email addresses to be copied on any correspondence about the case. The account that opens the case is already identified by passing the AWS Credentials in the HTTP POST method or in a method or function call from one of the programming languages supported by an AWS SDK .

A successful CreateCase request returns an AWS Support case number. Case numbers are used by the DescribeCases operation to retrieve existing AWS Support cases.

Creating a Request

createCase

Arguments

:: Text
ccSubject

-> Text
ccCommunicationBody

-> CreateCase

Creates a value of CreateCase with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccSeverityCode - The code for the severity level returned by the call to DescribeSeverityLevels .

ccIssueType - The type of issue for the case. You can specify either "customer-service" or "technical." If you do not indicate a value, the default is "technical."

ccCcEmailAddresses - A list of email addresses that AWS Support copies on case correspondence.

ccLanguage - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

ccCategoryCode - The category of problem for the AWS Support case.

ccServiceCode - The code for the AWS service returned by the call to DescribeServices .

ccAttachmentSetId - The ID of a set of one or more attachments for the case. Create the set by using AddAttachmentsToSet .

ccSubject - The title of the AWS Support case.

ccCommunicationBody - The communication body text when you create an AWS Support case by calling CreateCase .

data CreateCase

See: createCase smart constructor.

Request Lenses

ccSeverityCode :: Lens' CreateCase (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels .

ccIssueType :: Lens' CreateCase (Maybe Text)

The type of issue for the case. You can specify either "customer-service" or "technical." If you do not indicate a value, the default is "technical."

ccCcEmailAddresses :: Lens' CreateCase [Text]

A list of email addresses that AWS Support copies on case correspondence.

ccLanguage :: Lens' CreateCase (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

ccCategoryCode :: Lens' CreateCase (Maybe Text)

The category of problem for the AWS Support case.

ccServiceCode :: Lens' CreateCase (Maybe Text)

The code for the AWS service returned by the call to DescribeServices .

ccAttachmentSetId :: Lens' CreateCase (Maybe Text)

The ID of a set of one or more attachments for the case. Create the set by using AddAttachmentsToSet .

ccSubject :: Lens' CreateCase Text

The title of the AWS Support case.

ccCommunicationBody :: Lens' CreateCase Text

The communication body text when you create an AWS Support case by calling CreateCase .

Destructuring the Response

createCaseResponse

Arguments

:: Int
ccrsResponseStatus

-> CreateCaseResponse

Creates a value of CreateCaseResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccrsCaseId - The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

ccrsResponseStatus - -- | The response status code.

data CreateCaseResponse

The AWS Support case ID returned by a successful completion of the CreateCase operation.

See: createCaseResponse smart constructor.

Response Lenses

ccrsCaseId :: Lens' CreateCaseResponse (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

ccrsResponseStatus :: Lens' CreateCaseResponse Int

- | The response status code.