Network.AWS.Support.DescribeAttachment

Description

Returns the attachment that has the specified ID. Attachment IDs are generated by the case management system when you add an attachment to a case or case communication. Attachment IDs are returned in the AttachmentDetails objects that are returned by the DescribeCommunications operation.

Creating a Request

describeAttachment

Arguments

:: Text
daAttachmentId

-> DescribeAttachment

Creates a value of DescribeAttachment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daAttachmentId - The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

data DescribeAttachment

See: describeAttachment smart constructor.

Request Lenses

daAttachmentId :: Lens' DescribeAttachment Text

The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

Destructuring the Response

describeAttachmentResponse

Arguments

:: Int
darsResponseStatus

-> DescribeAttachmentResponse

Creates a value of DescribeAttachmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsAttachment - The attachment content and file name.

darsResponseStatus - -- | The response status code.

data DescribeAttachmentResponse

The content and file name of the attachment returned by the DescribeAttachment operation.

See: describeAttachmentResponse smart constructor.

Response Lenses

darsAttachment :: Lens' DescribeAttachmentResponse (Maybe Attachment)

The attachment content and file name.

darsResponseStatus :: Lens' DescribeAttachmentResponse Int

- | The response status code.