Network.AWS.Support.DescribeCommunications

Description

Returns communications (and attachments) for one or more support cases. You can use the afterTime and beforeTime parameters to filter by date. You can use the caseId parameter to restrict the results to a particular case.

Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error.

You can use the maxResults and nextToken parameters to control the pagination of the result set. Set maxResults to the number of cases you want displayed on each page, and use nextToken to specify the resumption of pagination.

This operation returns paginated results.

Creating a Request

describeCommunications

Arguments

:: Text
dCaseId

-> DescribeCommunications

Creates a value of DescribeCommunications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAfterTime - The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dBeforeTime - The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dNextToken - A resumption point for pagination.

dMaxResults - The maximum number of results to return before paginating.

dCaseId - The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

data DescribeCommunications

See: describeCommunications smart constructor.

Request Lenses

dAfterTime :: Lens' DescribeCommunications (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dBeforeTime :: Lens' DescribeCommunications (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dNextToken :: Lens' DescribeCommunications (Maybe Text)

A resumption point for pagination.

dMaxResults :: Lens' DescribeCommunications (Maybe Natural)

The maximum number of results to return before paginating.

dCaseId :: Lens' DescribeCommunications Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

Destructuring the Response

describeCommunicationsResponse

Arguments

:: Int
dcrsResponseStatus

-> DescribeCommunicationsResponse

Creates a value of DescribeCommunicationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrsNextToken - A resumption point for pagination.

dcrsCommunications - The communications for the case.

dcrsResponseStatus - -- | The response status code.

data DescribeCommunicationsResponse

The communications returned by the DescribeCommunications operation.

See: describeCommunicationsResponse smart constructor.

Response Lenses

dcrsNextToken :: Lens' DescribeCommunicationsResponse (Maybe Text)

A resumption point for pagination.

dcrsCommunications :: Lens' DescribeCommunicationsResponse [Communication]

The communications for the case.

dcrsResponseStatus :: Lens' DescribeCommunicationsResponse Int

- | The response status code.