Network.AWS.Support.Types

Service

data Support

Version 2013-04-15 of the Amazon Support service.

Error

data JSONError :: *

TrustedAdvisorResourcesSummary

data TrustedAdvisorResourcesSummary

trustedAdvisorResourcesSummary

Arguments

:: Integer
tarsResourcesProcessed

-> Integer
tarsResourcesFlagged

-> Integer
tarsResourcesIgnored

-> Integer
tarsResourcesSuppressed

-> TrustedAdvisorResourcesSummary

TrustedAdvisorResourcesSummary constructor.

The fields accessible through corresponding lenses are:

tarsResourcesFlagged :: Integer

tarsResourcesIgnored :: Integer

tarsResourcesProcessed :: Integer

tarsResourcesSuppressed :: Integer

tarsResourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

tarsResourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

tarsResourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

tarsResourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

TrustedAdvisorCategorySpecificSummary

data TrustedAdvisorCategorySpecificSummary

trustedAdvisorCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

TrustedAdvisorCategorySpecificSummary constructor.

The fields accessible through corresponding lenses are:

tacssCostOptimizing :: Maybe TrustedAdvisorCostOptimizingSummary

tacssCostOptimizing :: Lens' TrustedAdvisorCategorySpecificSummary (Maybe TrustedAdvisorCostOptimizingSummary)

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

Communication

data Communication

communication :: Communication

Communication constructor.

The fields accessible through corresponding lenses are:

cAttachmentSet :: [AttachmentDetails]

cBody :: Maybe Text

cCaseId :: Maybe Text

cSubmittedBy :: Maybe Text

cTimeCreated :: Maybe Text

cAttachmentSet :: Lens' Communication [AttachmentDetails]

Information about the attachments to the case communication.

cBody :: Lens' Communication (Maybe Text)

The text of the communication between the customer and AWS Support.

cCaseId :: Lens' Communication (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

cSubmittedBy :: Lens' Communication (Maybe Text)

The email address of the account that submitted the AWS Support case.

cTimeCreated :: Lens' Communication (Maybe Text)

The time the communication was created.

Category

data Category

category :: Category

Category constructor.

The fields accessible through corresponding lenses are:

cCode :: Maybe Text

cName :: Maybe Text

cCode :: Lens' Category (Maybe Text)

The category code for the support case.

cName :: Lens' Category (Maybe Text)

The category name for the support case.

TrustedAdvisorCheckSummary

data TrustedAdvisorCheckSummary

trustedAdvisorCheckSummary

Arguments

:: Text
tacsCheckId

-> Text
tacsTimestamp

-> Text
tacsStatus

-> TrustedAdvisorResourcesSummary
tacsResourcesSummary

-> TrustedAdvisorCategorySpecificSummary
tacsCategorySpecificSummary

-> TrustedAdvisorCheckSummary

TrustedAdvisorCheckSummary constructor.

The fields accessible through corresponding lenses are:

tacsCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

tacsCheckId :: Text

tacsHasFlaggedResources :: Maybe Bool

tacsResourcesSummary :: TrustedAdvisorResourcesSummary

tacsStatus :: Text

tacsTimestamp :: Text

tacsCategorySpecificSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

tacsCheckId :: Lens' TrustedAdvisorCheckSummary Text

The unique identifier for the Trusted Advisor check.

tacsHasFlaggedResources :: Lens' TrustedAdvisorCheckSummary (Maybe Bool)

Specifies whether the Trusted Advisor check has flagged resources.

tacsResourcesSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorResourcesSummary

tacsStatus :: Lens' TrustedAdvisorCheckSummary Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacsTimestamp :: Lens' TrustedAdvisorCheckSummary Text

The time of the last refresh of the check.

AttachmentDetails

data AttachmentDetails

attachmentDetails :: AttachmentDetails

AttachmentDetails constructor.

The fields accessible through corresponding lenses are:

adAttachmentId :: Maybe Text

adFileName :: Maybe Text

adAttachmentId :: Lens' AttachmentDetails (Maybe Text)

The ID of the attachment.

adFileName :: Lens' AttachmentDetails (Maybe Text)

The file name of the attachment.

TrustedAdvisorCheckResult

data TrustedAdvisorCheckResult

trustedAdvisorCheckResult

Arguments

:: Text
tacrCheckId

-> Text
tacrTimestamp

-> Text
tacrStatus

-> TrustedAdvisorResourcesSummary
tacrResourcesSummary

-> TrustedAdvisorCategorySpecificSummary
tacrCategorySpecificSummary

-> TrustedAdvisorCheckResult

TrustedAdvisorCheckResult constructor.

The fields accessible through corresponding lenses are:

tacrCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

tacrCheckId :: Text

tacrFlaggedResources :: [TrustedAdvisorResourceDetail]

tacrResourcesSummary :: TrustedAdvisorResourcesSummary

tacrStatus :: Text

tacrTimestamp :: Text

tacrCategorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

tacrCheckId :: Lens' TrustedAdvisorCheckResult Text

The unique identifier for the Trusted Advisor check.

tacrFlaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

tacrResourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

tacrStatus :: Lens' TrustedAdvisorCheckResult Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacrTimestamp :: Lens' TrustedAdvisorCheckResult Text

The time of the last refresh of the check.

TrustedAdvisorCheckDescription

data TrustedAdvisorCheckDescription

trustedAdvisorCheckDescription

Arguments

:: Text
tacdId

-> Text
tacdName

-> Text
tacdDescription

-> Text
tacdCategory

-> TrustedAdvisorCheckDescription

TrustedAdvisorCheckDescription constructor.

The fields accessible through corresponding lenses are:

tacdCategory :: Text

tacdDescription :: Text

tacdId :: Text

tacdMetadata :: [Text]

tacdName :: Text

tacdCategory :: Lens' TrustedAdvisorCheckDescription Text

The category of the Trusted Advisor check.

tacdDescription :: Lens' TrustedAdvisorCheckDescription Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended actions (contains HTML markup).

tacdId :: Lens' TrustedAdvisorCheckDescription Text

The unique identifier for the Trusted Advisor check.

tacdMetadata :: Lens' TrustedAdvisorCheckDescription [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

tacdName :: Lens' TrustedAdvisorCheckDescription Text

The display name for the Trusted Advisor check.

Attachment

data Attachment

attachment :: Attachment

Attachment constructor.

The fields accessible through corresponding lenses are:

aData :: Maybe Base64

aFileName :: Maybe Text

aData :: Lens' Attachment (Maybe Base64)

The content of the attachment file.

aFileName :: Lens' Attachment (Maybe Text)

The name of the attachment file.

RecentCaseCommunications

data RecentCaseCommunications

recentCaseCommunications :: RecentCaseCommunications

RecentCaseCommunications constructor.

The fields accessible through corresponding lenses are:

rccCommunications :: [Communication]

rccNextToken :: Maybe Text

rccCommunications :: Lens' RecentCaseCommunications [Communication]

The five most recent communications associated with the case.

rccNextToken :: Lens' RecentCaseCommunications (Maybe Text)

A resumption point for pagination.

SupportService

data SupportService

supportService :: SupportService

SupportService constructor.

The fields accessible through corresponding lenses are:

ssCategories :: [Category]

ssCode :: Maybe Text

ssName :: Maybe Text

ssCategories :: Lens' SupportService [Category]

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

ssCode :: Lens' SupportService (Maybe Text)

The code for an AWS service returned by the DescribeServices response. The Name element contains the corresponding friendly name.

ssName :: Lens' SupportService (Maybe Text)

The friendly name for an AWS service. The Code element contains the corresponding code.

TrustedAdvisorResourceDetail

data TrustedAdvisorResourceDetail

trustedAdvisorResourceDetail

Arguments

:: Text
tardStatus

-> Text
tardRegion

-> Text
tardResourceId

-> TrustedAdvisorResourceDetail

TrustedAdvisorResourceDetail constructor.

The fields accessible through corresponding lenses are:

tardIsSuppressed :: Maybe Bool

tardMetadata :: [Text]

tardRegion :: Text

tardResourceId :: Text

tardStatus :: Text

tardIsSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

tardMetadata :: Lens' TrustedAdvisorResourceDetail [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

tardRegion :: Lens' TrustedAdvisorResourceDetail Text

The AWS region in which the identified resource is located.

tardResourceId :: Lens' TrustedAdvisorResourceDetail Text

The unique identifier for the identified resource.

tardStatus :: Lens' TrustedAdvisorResourceDetail Text

The status code for the resource identified in the Trusted Advisor check.

TrustedAdvisorCostOptimizingSummary

data TrustedAdvisorCostOptimizingSummary

trustedAdvisorCostOptimizingSummary

Arguments

:: Double
tacosEstimatedMonthlySavings

-> Double
tacosEstimatedPercentMonthlySavings

-> TrustedAdvisorCostOptimizingSummary

TrustedAdvisorCostOptimizingSummary constructor.

The fields accessible through corresponding lenses are:

tacosEstimatedMonthlySavings :: Double

tacosEstimatedPercentMonthlySavings :: Double

tacosEstimatedMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated monthly savings that might be realized if the recommended actions are taken.

tacosEstimatedPercentMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated percentage of savings that might be realized if the recommended actions are taken.

SeverityLevel

data SeverityLevel

severityLevel :: SeverityLevel

SeverityLevel constructor.

The fields accessible through corresponding lenses are:

slCode :: Maybe Text

slName :: Maybe Text

slCode :: Lens' SeverityLevel (Maybe Text)

One of four values: "low," "medium," "high," and "urgent". These values correspond to response times returned to the caller in name.

slName :: Lens' SeverityLevel (Maybe Text)

The name of the severity level that corresponds to the severity level code.

CaseDetails

data CaseDetails

caseDetails :: CaseDetails

CaseDetails constructor.

The fields accessible through corresponding lenses are:

cdCaseId :: Maybe Text

cdCategoryCode :: Maybe Text

cdCcEmailAddresses :: [Text]

cdDisplayId :: Maybe Text

cdLanguage :: Maybe Text

cdRecentCommunications :: Maybe RecentCaseCommunications

cdServiceCode :: Maybe Text

cdSeverityCode :: Maybe Text

cdStatus :: Maybe Text

cdSubject :: Maybe Text

cdSubmittedBy :: Maybe Text

cdTimeCreated :: Maybe Text

cdCaseId :: Lens' CaseDetails (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

cdCategoryCode :: Lens' CaseDetails (Maybe Text)

The category of problem for the AWS Support case.

cdCcEmailAddresses :: Lens' CaseDetails [Text]

The email addresses that receive copies of communication about the case.

cdDisplayId :: Lens' CaseDetails (Maybe Text)

The ID displayed for the case in the AWS Support Center. This is a numeric string.

cdLanguage :: Lens' CaseDetails (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

cdRecentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

cdServiceCode :: Lens' CaseDetails (Maybe Text)

The code for the AWS service returned by the call to DescribeServices.

cdSeverityCode :: Lens' CaseDetails (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels .

cdStatus :: Lens' CaseDetails (Maybe Text)

The status of the case.

cdSubject :: Lens' CaseDetails (Maybe Text)

The subject line for the case in the AWS Support Center.

cdSubmittedBy :: Lens' CaseDetails (Maybe Text)

The email address of the account that submitted the case.

cdTimeCreated :: Lens' CaseDetails (Maybe Text)

The time that the case was case created in the AWS Support Center.

TrustedAdvisorCheckRefreshStatus

data TrustedAdvisorCheckRefreshStatus

trustedAdvisorCheckRefreshStatus

Arguments

:: Text
tacrsCheckId

-> Text
tacrsStatus

-> Integer
tacrsMillisUntilNextRefreshable

-> TrustedAdvisorCheckRefreshStatus

TrustedAdvisorCheckRefreshStatus constructor.

The fields accessible through corresponding lenses are:

tacrsCheckId :: Text

tacrsMillisUntilNextRefreshable :: Integer

tacrsStatus :: Text

tacrsCheckId :: Lens' TrustedAdvisorCheckRefreshStatus Text

The unique identifier for the Trusted Advisor check.

tacrsMillisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

tacrsStatus :: Lens' TrustedAdvisorCheckRefreshStatus Text

The status of the Trusted Advisor check for which a refresh has been requested: "none", "enqueued", "processing", "success", or "abandoned".