Network.AWS.Support.V2013_04_15.DescribeCommunications

Description

Returns communications (and attachments) for one or more support cases. You can use the AfterTime and BeforeTime parameters to filter by date. You can use the CaseId parameter to restrict the results to a particular case. Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error. You can use the MaxResults and NextToken parameters to control the pagination of the result set. Set MaxResults to the number of cases you want displayed on each page, and use NextToken to specify the resumption of pagination.

Request

data DescribeCommunications

Request constructor

mkDescribeCommunications

Arguments

:: Text
dc1CaseId

-> DescribeCommunications

Smart constructor for the minimum required parameters to construct a valid DescribeCommunications request.

The fields accessible through corresponding lenses are:

CaseId :: Text

BeforeTime :: Maybe Text

AfterTime :: Maybe Text

NextToken :: Maybe Text

MaxResults :: Maybe Integer

Request lenses

dc1CaseId :: Lens' DescribeCommunications Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

dc1BeforeTime :: Lens' DescribeCommunications (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dc1AfterTime :: Lens' DescribeCommunications (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

dc1NextToken :: Lens' DescribeCommunications (Maybe Text)

A resumption point for pagination.

dc1MaxResults :: Lens' DescribeCommunications (Maybe Integer)

The maximum number of results to return before paginating.

Response

data DescribeCommunicationsResponse

The communications returned by the DescribeCommunications operation.

Response constructor

mkDescribeCommunicationsResponse :: DescribeCommunicationsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeCommunicationsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Communications :: [Communication]

NextToken :: Maybe Text

Response lenses

dcrrCommunications :: Lens' DescribeCommunicationsResponse [Communication]

The communications for the case.

dcrrNextToken :: Lens' DescribeCommunicationsResponse (Maybe Text)

A resumption point for pagination.