Network.AWS.Support.V2013_04_15.DescribeSeverityLevels

Description

Returns the list of severity levels that you can assign to an AWS Support case. The severity level for a case is also a field in the CaseDetails data type included in any CreateCase request.

Documentation

describeSeverityLevels :: DescribeSeverityLevels

Minimum specification for a DescribeSeverityLevels request.

data DescribeSeverityLevels

Constructors

DescribeSeverityLevels

Fields

_dslrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

dslrLanguage :: Iso' DescribeSeverityLevels (Maybe Text)

data DescribeSeverityLevelsResponse

Constructors

DescribeSeverityLevelsResponse

Fields

_dslsSeverityLevels :: [SeverityLevel]

The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

dslsSeverityLevels :: Iso' DescribeSeverityLevelsResponse [SeverityLevel]