Network.AWS.Support.V2013_04_15.DescribeTrustedAdvisorChecks

Description

Returns information about all available Trusted Advisor checks, including name, ID, category, description, and metadata. You must specify a language code; English ("en") and Japanese ("ja") are currently supported. The response contains a TrustedAdvisorCheckDescription for each check.

Documentation

data DescribeTrustedAdvisorChecks

Constructors

DescribeTrustedAdvisorChecks

Fields

_dtacrLanguage :: Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

dtacrLanguage :: Iso' DescribeTrustedAdvisorChecks Text

data DescribeTrustedAdvisorChecksResponse

Constructors

DescribeTrustedAdvisorChecksResponse

Fields

_dtacsChecks :: [TrustedAdvisorCheckDescription]

Information about all available Trusted Advisor checks.

dtacsChecks :: Iso' DescribeTrustedAdvisorChecksResponse [TrustedAdvisorCheckDescription]