Network.AWS.Support.V2013_04_15.RefreshTrustedAdvisorCheck

Description

Requests a refresh of the Trusted Advisor check that has the specified check ID. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks. The response contains a RefreshTrustedAdvisorCheckResult object, which contains these fields: Status. The refresh status of the check: "none", "enqueued", "processing", "success", or "abandoned". MillisUntilNextRefreshable. The amount of time, in milliseconds, until the check is eligible for refresh. CheckId. The unique identifier for the check.

Request

data RefreshTrustedAdvisorCheck

Request constructor

mkRefreshTrustedAdvisorCheck

Arguments

:: Text
rtacCheckId

-> RefreshTrustedAdvisorCheck

Smart constructor for the minimum required parameters to construct a valid RefreshTrustedAdvisorCheck request.

Request lenses

rtacCheckId :: Lens' RefreshTrustedAdvisorCheck Text

The unique identifier for the Trusted Advisor check.

Response

data RefreshTrustedAdvisorCheckResponse

The current refresh status of a Trusted Advisor check.

Response constructor

mkRefreshTrustedAdvisorCheckResponse

Arguments

:: TrustedAdvisorCheckRefreshStatus
rtacrStatus

-> RefreshTrustedAdvisorCheckResponse

Smart constructor for the minimum required parameters to construct a valid RefreshTrustedAdvisorCheckResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

rtacrStatus :: Lens' RefreshTrustedAdvisorCheckResponse TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.