Network.AWS.Support.V2013_04_15.Types

Description

AWS Support is a one-on-one, fast-response support channel that is staffed with experienced support personnel. The service helps customers get the most from the products and features provided by Amazon Web Services. There are four levels, or tiers, of AWS Support: Basic, Developer, Business, and Enterprise. The Basic tier is free of charge and offers support for account and billing questions and service limit increases. The other tiers offer an unlimited number of technical support cases with pay-by-the-month pricing and no long-term contracts, providing developers and businesses flexibility to choose the level of support that meets their needs.

Service

data Support

Supported version (2013-04-15) of the AWS Support service.

TrustedAdvisorCategorySpecificSummary

data TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

mkTrustedAdvisorCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorCategorySpecificSummary data type to populate a request.

tacssCostOptimizing :: Lens' TrustedAdvisorCategorySpecificSummary (Maybe TrustedAdvisorCostOptimizingSummary)

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

Attachment

data Attachment

An attachment to a case communication. The attachment consists of the file name and the content of the file.

mkAttachment :: Attachment

Smart constructor for the minimum required fields to construct a valid Attachment data type to populate a request.

aFileName :: Lens' Attachment (Maybe Text)

The name of the attachment file.

aData :: Lens' Attachment (Maybe Base64)

The content of the attachment file.

AttachmentDetails

data AttachmentDetails

The file name and ID of an attachment to a case communication. You can use the ID to retrieve the attachment with the DescribeAttachment operation.

mkAttachmentDetails :: AttachmentDetails

Smart constructor for the minimum required fields to construct a valid AttachmentDetails data type to populate a request.

adAttachmentId :: Lens' AttachmentDetails (Maybe Text)

The ID of the attachment.

adFileName :: Lens' AttachmentDetails (Maybe Text)

The file name of the attachment.

CaseDetails

data CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained the response from a DescribeCases request. CaseDetails contains the following fields: CaseID. The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47. CategoryCode. The category of problem for the AWS Support case. Corresponds to the CategoryCode values returned by a call to DescribeServices. DisplayId. The identifier for the case on pages in the AWS Support Center. Language. The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them. RecentCommunications. One or more Communication objects. Fields of these objects are Attachments, Body, CaseId, SubmittedBy, and TimeCreated. NextToken. A resumption point for pagination. ServiceCode. The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices. SeverityCode. The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels. Status. The status of the case in the AWS Support Center. Subject. The subject line of the case. SubmittedBy. The email address of the account that submitted the case. TimeCreated. The time the case was created, in ISO-8601 format.

mkCaseDetails :: CaseDetails

Smart constructor for the minimum required fields to construct a valid CaseDetails data type.

CaseDetails is exclusively used in responses and this constructor is provided for convenience and testing purposes.

cdCaseId :: Lens' CaseDetails (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

cdDisplayId :: Lens' CaseDetails (Maybe Text)

The ID displayed for the case in the AWS Support Center. This is a numeric string.

cdSubject :: Lens' CaseDetails (Maybe Text)

The subject line for the case in the AWS Support Center.

cdStatus :: Lens' CaseDetails (Maybe Text)

The status of the case.

cdServiceCode :: Lens' CaseDetails (Maybe Text)

The code for the AWS service returned by the call to DescribeServices.

cdCategoryCode :: Lens' CaseDetails (Maybe Text)

The category of problem for the AWS Support case.

cdSeverityCode :: Lens' CaseDetails (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels.

cdSubmittedBy :: Lens' CaseDetails (Maybe Text)

The email address of the account that submitted the case.

cdTimeCreated :: Lens' CaseDetails (Maybe Text)

The time that the case was case created in the AWS Support Center.

cdRecentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

cdCcEmailAddresses :: Lens' CaseDetails [Text]

The email addresses that receive copies of communication about the case.

cdLanguage :: Lens' CaseDetails (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Category

data Category

A JSON-formatted name/value pair that represents the category name and category code of the problem, selected from the DescribeServices response for each AWS service.

mkCategory :: Category

Smart constructor for the minimum required fields to construct a valid Category data type to populate a request.

c1Code :: Lens' Category (Maybe Text)

The category code for the support case.

c1Name :: Lens' Category (Maybe Text)

The category name for the support case.

Communication

data Communication

A communication associated with an AWS Support case. The communication consists of the case ID, the message body, attachment information, the account email address, and the date and time of the communication.

mkCommunication :: Communication

Smart constructor for the minimum required fields to construct a valid Communication data type to populate a request.

cCaseId :: Lens' Communication (Maybe Text)

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

cBody :: Lens' Communication (Maybe Text)

The text of the communication between the customer and AWS Support.

cSubmittedBy :: Lens' Communication (Maybe Text)

The email address of the account that submitted the AWS Support case.

cTimeCreated :: Lens' Communication (Maybe Text)

The time the communication was created.

cAttachmentSet :: Lens' Communication [AttachmentDetails]

Information about the attachments to the case communication.

RecentCaseCommunications

data RecentCaseCommunications

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

mkRecentCaseCommunications :: RecentCaseCommunications

Smart constructor for the minimum required fields to construct a valid RecentCaseCommunications data type to populate a request.

rccCommunications :: Lens' RecentCaseCommunications [Communication]

The five most recent communications associated with the case.

rccNextToken :: Lens' RecentCaseCommunications (Maybe Text)

A resumption point for pagination.

Service

data Service

Information about an AWS service returned by the DescribeServices operation.

mkService :: Service

Smart constructor for the minimum required fields to construct a valid Service data type.

Service is exclusively used in responses and this constructor is provided for convenience and testing purposes.

sCode :: Lens' Service (Maybe Text)

The code for an AWS service returned by the DescribeServices response. The Name element contains the corresponding friendly name.

sName :: Lens' Service (Maybe Text)

The friendly name for an AWS service. The Code element contains the corresponding code.

sCategories :: Lens' Service [Category]

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

SeverityLevel

data SeverityLevel

A code and name pair that represent a severity level that can be applied to a support case.

mkSeverityLevel :: SeverityLevel

Smart constructor for the minimum required fields to construct a valid SeverityLevel data type.

SeverityLevel is exclusively used in responses and this constructor is provided for convenience and testing purposes.

slCode :: Lens' SeverityLevel (Maybe Text)

One of four values: "low," "medium," "high," and "urgent". These values correspond to response times returned to the caller in SeverityLevel.name.

slName :: Lens' SeverityLevel (Maybe Text)

The name of the severity level that corresponds to the severity level code.

TrustedAdvisorCheckDescription

data TrustedAdvisorCheckDescription

The description and metadata for a Trusted Advisor check.

mkTrustedAdvisorCheckDescription

Arguments

:: Text
tacdId

-> Text
tacdName

-> Text
tacdDescription

-> Text
tacdCategory

-> [Text]
tacdMetadata

-> TrustedAdvisorCheckDescription

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorCheckDescription data type.

TrustedAdvisorCheckDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

tacdId :: Lens' TrustedAdvisorCheckDescription Text

The unique identifier for the Trusted Advisor check.

tacdName :: Lens' TrustedAdvisorCheckDescription Text

The display name for the Trusted Advisor check.

tacdDescription :: Lens' TrustedAdvisorCheckDescription Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended actions (contains HTML markup).

tacdCategory :: Lens' TrustedAdvisorCheckDescription Text

The category of the Trusted Advisor check.

tacdMetadata :: Lens' TrustedAdvisorCheckDescription [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorCheckRefreshStatus

data TrustedAdvisorCheckRefreshStatus

The refresh status of a Trusted Advisor check.

mkTrustedAdvisorCheckRefreshStatus

Arguments

:: Text
tacrsCheckId

-> Text
tacrsStatus

-> Integer
tacrsMillisUntilNextRefreshable

-> TrustedAdvisorCheckRefreshStatus

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorCheckRefreshStatus data type.

TrustedAdvisorCheckRefreshStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

tacrsCheckId :: Lens' TrustedAdvisorCheckRefreshStatus Text

The unique identifier for the Trusted Advisor check.

tacrsStatus :: Lens' TrustedAdvisorCheckRefreshStatus Text

The status of the Trusted Advisor check for which a refresh has been requested: "none", "enqueued", "processing", "success", or "abandoned".

tacrsMillisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

TrustedAdvisorCheckResult

data TrustedAdvisorCheckResult

The detailed results of the Trusted Advisor check.

mkTrustedAdvisorCheckResult

Arguments

:: Text
tacrCheckId

-> Text
tacrTimestamp

-> Text
tacrStatus

-> TrustedAdvisorResourcesSummary
tacrResourcesSummary

-> TrustedAdvisorCategorySpecificSummary
tacrCategorySpecificSummary

-> [TrustedAdvisorResourceDetail]
tacrFlaggedResources

-> TrustedAdvisorCheckResult

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorCheckResult data type.

TrustedAdvisorCheckResult is exclusively used in responses and this constructor is provided for convenience and testing purposes.

tacrCheckId :: Lens' TrustedAdvisorCheckResult Text

The unique identifier for the Trusted Advisor check.

tacrTimestamp :: Lens' TrustedAdvisorCheckResult Text

The time of the last refresh of the check.

tacrStatus :: Lens' TrustedAdvisorCheckResult Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacrResourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

tacrCategorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

tacrFlaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

TrustedAdvisorCheckSummary

data TrustedAdvisorCheckSummary

A summary of a Trusted Advisor check result, including the alert status, last refresh, and number of resources examined.

mkTrustedAdvisorCheckSummary

Arguments

:: Text
tacsCheckId

-> Text
tacsTimestamp

-> Text
tacsStatus

-> TrustedAdvisorResourcesSummary
tacsResourcesSummary

-> TrustedAdvisorCategorySpecificSummary
tacsCategorySpecificSummary

-> TrustedAdvisorCheckSummary

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorCheckSummary data type.

TrustedAdvisorCheckSummary is exclusively used in responses and this constructor is provided for convenience and testing purposes.

tacsCheckId :: Lens' TrustedAdvisorCheckSummary Text

The unique identifier for the Trusted Advisor check.

tacsTimestamp :: Lens' TrustedAdvisorCheckSummary Text

The time of the last refresh of the check.

tacsStatus :: Lens' TrustedAdvisorCheckSummary Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

tacsHasFlaggedResources :: Lens' TrustedAdvisorCheckSummary (Maybe Bool)

Specifies whether the Trusted Advisor check has flagged resources.

tacsResourcesSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

tacsCategorySpecificSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

TrustedAdvisorCostOptimizingSummary

data TrustedAdvisorCostOptimizingSummary

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

mkTrustedAdvisorCostOptimizingSummary

Arguments

:: Double
tacosEstimatedMonthlySavings

-> Double
tacosEstimatedPercentMonthlySavings

-> TrustedAdvisorCostOptimizingSummary

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorCostOptimizingSummary data type to populate a request.

tacosEstimatedMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated monthly savings that might be realized if the recommended actions are taken.

tacosEstimatedPercentMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated percentage of savings that might be realized if the recommended actions are taken.

TrustedAdvisorResourceDetail

data TrustedAdvisorResourceDetail

Contains information about a resource identified by a Trusted Advisor check.

mkTrustedAdvisorResourceDetail

Arguments

:: Text
tardStatus

-> Text
tardRegion

-> Text
tardResourceId

-> [Text]
tardMetadata

-> TrustedAdvisorResourceDetail

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorResourceDetail data type.

TrustedAdvisorResourceDetail is exclusively used in responses and this constructor is provided for convenience and testing purposes.

tardStatus :: Lens' TrustedAdvisorResourceDetail Text

The status code for the resource identified in the Trusted Advisor check.

tardRegion :: Lens' TrustedAdvisorResourceDetail Text

The AWS region in which the identified resource is located.

tardResourceId :: Lens' TrustedAdvisorResourceDetail Text

The unique identifier for the identified resource.

tardIsSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

tardMetadata :: Lens' TrustedAdvisorResourceDetail [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorResourcesSummary

data TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

mkTrustedAdvisorResourcesSummary

Arguments

:: Integer
tarsResourcesProcessed

-> Integer
tarsResourcesFlagged

-> Integer
tarsResourcesIgnored

-> Integer
tarsResourcesSuppressed

-> TrustedAdvisorResourcesSummary

Smart constructor for the minimum required fields to construct a valid TrustedAdvisorResourcesSummary data type to populate a request.

tarsResourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

tarsResourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

tarsResourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

tarsResourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.